% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/add_edge_df.R
\name{add_edge_df}
\alias{add_edge_df}
\title{Add edges from an edge data frame to an existing graph object}
\usage{
add_edge_df(graph, edge_df)
}
\arguments{
\item{graph}{a graph object of class \code{dgr_graph} that is created using
\code{create_graph}.}

\item{edge_df}{an edge data frame that is created using \code{create_edges}.}
}
\value{
a graph object of class \code{dgr_graph}.
}
\description{
With a graph object of class \code{dgr_graph} add edges from an
edge data frame to that graph.
}
\examples{
\dontrun{
}
}

