% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_edges_from_csv.R
\name{add_edges_from_csv}
\alias{add_edges_from_csv}
\title{Add edges and attributes from a CSV file}
\usage{
add_edges_from_csv(graph, csv_path, from_col, from_attr, to_col, to_attr,
  set_rel = NULL, select_cols = NULL, drop_cols = NULL,
  rename_attrs = NULL, rel_col = NULL)
}
\arguments{
\item{graph}{a graph object of class \code{dgr_graph} that is created
using \code{create_graph}.}

\item{csv_path}{a path to a CSV file.}

\item{from_col}{from_col}

\item{from_attr}{from_attr}

\item{to_col}{to_col}

\item{to_attr}{to_attr}

\item{set_rel}{an optional string to apply a \code{rel} attribute to
all edges created from the CSV records.}

\item{select_cols}{an optional character vector for specifying which
columns in the CSV file should be imported as edge attributes.}

\item{rename_attrs}{an optional character for renaming edge attributes.}

\item{rel_col}{an option to apply a column of data in the CSV file
as \code{rel} attribute values.}
}
\value{
a graph object of class \code{dgr_graph}.
}
\description{
Add nodes and their attributes to an existing graph object
from data in a CSV file.
}

