% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_edges_from_table.R
\name{add_edges_from_table}
\alias{add_edges_from_table}
\title{Add edges and attributes to graph from a table}
\usage{
add_edges_from_table(graph, table, from_col, to_col, ndf_mapping,
  set_rel = NULL, select_cols = NULL, drop_cols = NULL, rel_col = NULL)
}
\arguments{
\item{graph}{a graph object of class
\code{dgr_graph} that is created using
\code{create_graph}.}

\item{table}{either a path to a CSV file, or, a data
frame object.}

\item{from_col}{the name of the table column from
which edges originate.}

\item{to_col}{the name of the table column to
which edges terminate.}

\item{ndf_mapping}{a single character value for
the mapping of the \code{from} and \code{to} columns
in the external table (supplied as \code{from_col}
and \code{to_col}, respectively) to a column in the
graph's internal node data frame (ndf).}

\item{set_rel}{an optional string to apply a
\code{rel} attribute to all edges created from the
table records.}

\item{select_cols}{an optional character vector for
specifying which columns in the table that should be
imported as edge attributes.}

\item{drop_cols}{an optional character vector for
dropping columns from the incoming data.}

\item{rel_col}{an option to apply a column of data
in the table as \code{rel} attribute values.}
}
\value{
a graph object of class \code{dgr_graph}.
}
\description{
Add edges and their attributes to an
existing graph object from data in a CSV file or a
data frame.
}
\examples{
\dontrun{
# Create a graph from a CSV file
graph <-
  create_graph() \%>\%
  add_edges_from_table(
    system.file("examples/projects_and_contributors.csv",
                package = "DiagrammeR"),
    from_col = "contributor_name",
    to_col = "project_name",
    rel_col = "contributor_role",
    set_rel = "contributes_to")

# Get a summary of the new graph
graph_info(graph)
#>   name  n  e      dens min_deg max_deg avg_deg time tz
#> 1      13 13 0.1666667       1       7       2
}
}

