% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_full_graph.R
\name{add_full_graph}
\alias{add_full_graph}
\title{Add a fully connected graph}
\usage{
add_full_graph(graph, n, type = NULL, label = TRUE, rel = NULL,
  edge_wt_matrix = NULL, keep_loops = FALSE)
}
\arguments{
\item{graph}{a graph object of class
\code{dgr_graph} that is created using
\code{create_graph}.}

\item{n}{the number of nodes comprising the fully
connected graph.}

\item{type}{an optional string that describes the
entity type for the nodes to be added.}

\item{label}{either a vector object of length
\code{n} that provides optional labels for the new
nodes, or, a boolean value where setting to
\code{TRUE} ascribes node IDs to the label and
\code{FALSE} or \code{NULL} yields a blank label.}

\item{rel}{an optional string for providing a
relationship label to all new edges created in the
connected graph.}

\item{edge_wt_matrix}{an optional matrix of \code{n}
by \code{n} dimensions containing values to apply
as edge weights.}
}
\value{
a graph object of class \code{dgr_graph}.
}
\description{
With a graph object of class
\code{dgr_graph}, add a fully connected graph either
with or without loops. If the graph object set as
directed, the added graph will have edges to and from
each pair of nodes. In the undirected case, a single
edge will link each pair of nodes.
}

