% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_growing_graph.R
\name{add_growing_graph}
\alias{add_growing_graph}
\title{Create a random growing graph with m edges added per step}
\usage{
add_growing_graph(graph, n, m = 1, citation = FALSE, set_seed = NULL)
}
\arguments{
\item{graph}{a graph object of class
\code{dgr_graph}.}

\item{n}{the number of nodes comprising the
generated graph.}

\item{m}{the number of edges added per time
step.}

\item{citation}{a logical value (default is
\code{FALSE}) that governs whether a citation
graph is to be created. This is where new
edges specifically originate from the newly
added node in the most recent time step.}

\item{set_seed}{supplying a value sets a random seed
of the \code{Mersenne-Twister} implementation.}
}
\description{
To an existing graph object, add
a graph built by adding \code{m} new edges
at each time step (where a node is added).
}
\examples{
# Create a random, growing
# graph with 100 nodes
growing_graph <-
  create_graph(
    directed = FALSE) \%>\%
  add_growing_graph(
    n = 100,
    citation = TRUE,
    set_seed = 23)

# Get a count of nodes
growing_graph \%>\%
  node_count()
#> [1] 100

# Get a count of edges
growing_graph \%>\%
  edge_count()
#> [1] 99
}
