% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_islands_graph.R
\name{add_islands_graph}
\alias{add_islands_graph}
\title{Create a random growing graph with m edges between islands}
\usage{
add_islands_graph(graph, n_islands, island_size, p, edges_between,
  set_seed = NULL)
}
\arguments{
\item{graph}{a graph object of class
\code{dgr_graph}.}

\item{n_islands}{the number of islands
in the generated graph.}

\item{island_size}{the size of the islands
in the generated graph.}

\item{p}{the probability of there being
edges between the islands.}

\item{edges_between}{The number of edges
between islands.}

\item{set_seed}{supplying a value sets a random seed
of the \code{Mersenne-Twister} implementation.}
}
\description{
To an existing graph object, add
several Erdos-Renyi random graphs (the islands)
using a common set of parameters, connected
together by a fixed number of edges.
}
\examples{
# Create a graph of islands
islands_graph <-
  create_graph() \%>\%
  add_islands_graph(
    n_islands = 4,
    island_size = 10,
    p = 0.5,
    edges_between = 1,
    set_seed = 23)

# Get a count of nodes
islands_graph \%>\%
  count_nodes()
#> [1] 40

# Get a count of edges
islands_graph \%>\%
  count_edges()
#> [1] 98
}
