% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grViz.R
\name{add_mathjax}
\alias{add_mathjax}
\title{Add MathJax-formatted equation text}
\usage{
add_mathjax(gv = NULL, include_mathjax = TRUE)
}
\arguments{
\item{gv}{a \code{grViz} htmlwidget.}

\item{include_mathjax}{\code{logical} to add mathjax JS.
Change to \code{FALSE} if using with \code{RMarkdown} since
MathJax will likely already be added.}
}
\value{
a \code{grViz} htmlwidget
}
\description{
Add MathJax-formatted equation text
}
