% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/add_node.R
\name{add_node}
\alias{add_node}
\title{Add a node to an existing graph object}
\usage{
add_node(graph, node = NULL, from = NULL, to = NULL, label = NULL,
  type = NULL)
}
\arguments{
\item{graph}{a graph object of class 'gv_graph' that is created using 'graphviz_graph'.}

\item{node}{a node ID for the newly connected node.}

\item{from}{a vector containing node IDs from which edges will be directed to the new node.}

\item{to}{a vector containing node IDs to which edges will be directed from the new node.}

\item{label}{an optional label to be ascribed to the node for visual display.}

\item{type}{a vector of strings reppresenting the entity type for the nodes.}
}
\value{
a graph object of class 'gv_graph'.
}
\description{
With a graph object of class 'gv_graph', add a new node of a specified type to extant nodes within the graph.
}

