% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_nodes_from_df_cols.R
\name{add_nodes_from_df_cols}
\alias{add_nodes_from_df_cols}
\title{Add nodes from distinct values in data frame columns}
\usage{
add_nodes_from_df_cols(graph, df, columns, type = NULL)
}
\arguments{
\item{graph}{a graph object of class
\code{dgr_graph}.}

\item{df}{a data frame from which values will be
taken as new nodes for the graph.}

\item{columns}{a character vector of column names
or a numeric vector of column numbers for the
data frame supplied in \code{df}. The distinct
values in these columns will serve as labels for
the nodes added to the graph.}

\item{type}{an optional, single-length character
vector that provides a group identifier for the
nodes to be added to the graph.}
}
\value{
a graph object of class \code{dgr_graph}.
}
\description{
Add new nodes to a graph object of
class \code{dgr_graph} using distinct values from
one or more columns in a data frame. The values will
serve as node labels and the number of nodes added
depends on the number of distinct values found in
the specified columns.
}
\examples{
# Create a simple graph
graph <-
  create_graph() \%>\%
  add_path(2)

# Create a data frame from which several
# columns have values designated as graph nodes
df <-
  data.frame(
    col_1 = c("f", "p", "q"),
    col_2 = c("q", "x", "f"),
    col_3 = c(1, 5, 3),
    col_4 = c("a", "v", "h"),
    stringsAsFactors = FALSE)

# Add nodes from columns `col_1` and `col_2`
# from the data frame to the graph object
graph <-
  graph \%>\%
  add_nodes_from_df_cols(
    df = df,
    columns = c("col_1", "col_2"))

# Show the graph's node data frame
graph \%>\% get_node_df()
#>   id type label
#> 1  1 <NA>     1
#> 2  2 <NA>     2
#> 3  3 <NA>     f
#> 4  4 <NA>     p
#> 5  5 <NA>     q
#> 6  6 <NA>     x
}

