% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/add_to_graph_series.R
\name{add_to_graph_series}
\alias{add_to_graph_series}
\title{Add graph object to a graph series object
Add a graph object to an extant graph series object for storage of multiple graphs across a sequential or temporal one-dimensional array.}
\usage{
add_to_graph_series(graph, graph_series)
}
\arguments{
\item{graph}{a graph object to add to the graph series object.}

\item{graph_series}{a graph series object to which the graph object will be added.}
}
\value{
a graph series object of type \code{dgr_graph_1D}.
}
\description{
Add graph object to a graph series object
Add a graph object to an extant graph series object for storage of multiple graphs across a sequential or temporal one-dimensional array.
}

