% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_node_attrs.R
\name{cache_node_attrs}
\alias{cache_node_attrs}
\title{Cache node attributes in the graph}
\usage{
cache_node_attrs(graph, node_attr, mode = NULL, nodes = NULL)
}
\arguments{
\item{graph}{a graph object of class
\code{dgr_graph}.}

\item{node_attr}{the node attribute from which to
obtain values.}

\item{mode}{a option to recast the returned vector
of node attribute value as \code{numeric} or
\code{character}.}

\item{nodes}{an optional vector of node IDs for
filtering the list of nodes present in the graph.}
}
\value{
a graph object of class \code{dgr_graph}.
}
\description{
From a graph object of class
\code{dgr_graph}, get node attribute properties for
one or more nodes and cache those values in the
graph for later retrieval using \code{get_cache}.
}

