% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_node_count_ws.R
\name{cache_node_count_ws}
\alias{cache_node_count_ws}
\title{Cache a count of nodes (available in a selection)
in the graph}
\usage{
cache_node_count_ws(graph)
}
\arguments{
\item{graph}{a graph object of class
\code{dgr_graph} that is created using
\code{create_graph}.}
}
\value{
a graph object of class \code{dgr_graph}.
}
\description{
From a graph object of class
\code{dgr_graph}, get a count of nodes available in
a selection and cache that value in the graph for
later retrieval using \code{get_cache}.
}
\examples{
# Create a graph with 10 nodes and 9 edges
graph <-
  create_graph() \%>\%
  add_n_nodes(10) \%>\%
  add_edges_w_string(
    "1->2 1->3 2->4 2->5 3->6 3->7 4->8 4->9 5->10")

# Cache a count of nodes after creating a selection
graph <-
  graph \%>\%
  select_nodes_by_id(2:5) \%>\%
  cache_node_count_ws()

# Get the number of nodes stored in the cache
graph \%>\% get_cache
#> [1] 4
}

