% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_ndfs.R
\name{combine_ndfs}
\alias{combine_ndfs}
\title{Combine multiple node data frames}
\usage{
combine_ndfs(...)
}
\arguments{
\item{...}{two or more node data frames, which
contain node IDs and associated attributes.}
}
\value{
a combined node data frame.
}
\description{
Combine several node data frames
into a single node data frame.
}

