% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_nodes.R
\name{combine_nodes}
\alias{combine_nodes}
\title{Combine multiple node data frames into a single
node data frame}
\usage{
combine_nodes(...)
}
\arguments{
\item{...}{two or more node data frames, which
contain node IDs and associated attributes.}
}
\value{
a combined node data frame.
}
\description{
Combine several node data frames in the
style of \code{rbind}, except, it works regardless
of the number and ordering of the columns.
}
\examples{
\dontrun{
# Create a node data frame (ndf)
nodes_1 <-
  create_nodes(
    nodes = 1:4,
    label = FALSE,
    type = "lower",
    style = "filled",
    color = "aqua",
    shape = c("circle", "circle",
              "rectangle", "rectangle"),
    data = c(3.5, 2.6, 9.4, 2.7))

# View the `nodes_1` ndf
nodes_1
#>   nodes  type label  style color     shape data
#> 1     1 lower       filled  aqua    circle  3.5
#> 2     2 lower       filled  aqua    circle  2.6
#> 3     3 lower       filled  aqua rectangle  9.4
#> 4     4 lower       filled  aqua rectangle  2.7

# Create another node data frame
nodes_2 <-
  create_nodes(
    nodes = 5:8,
    label = FALSE,
    type = "upper",
    style = "filled",
    color = "red",
    shape = "triangle",
    data = c(0.5, 3.9, 3.7, 8.2))

# View the `nodes_2` ndf
nodes_2
#>   nodes  type label  style color    shape data
#> 1     5 upper       filled   red triangle  0.5
#> 2     6 upper       filled   red triangle  3.9
#> 3     7 upper       filled   red triangle  3.7
#> 4     8 upper       filled   red triangle  8.2

# Combine the two node data frames
all_nodes <-
  combine_nodes(nodes_1, nodes_2)

# View the combined node data frame
all_nodes
#>   nodes  type label  style color     shape data
#> 1     1 lower       filled  aqua    circle  3.5
#> 2     2 lower       filled  aqua    circle  2.6
#> 3     3 lower       filled  aqua rectangle  9.4
#> 4     4 lower       filled  aqua rectangle  2.7
#> 5     5 upper       filled   red  triangle  0.5
#> 6     6 upper       filled   red  triangle  3.9
#> 7     7 upper       filled   red  triangle  3.7
#> 8     8 upper       filled   red  triangle  8.2
}
}

