% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_edge_attrs.R
\name{copy_edge_attrs}
\alias{copy_edge_attrs}
\title{Copy an edge attribute column and set the name}
\usage{
copy_edge_attrs(graph, edge_attr_from, edge_attr_to)
}
\arguments{
\item{graph}{a graph object of class}

\item{edge_attr_from}{the name of the edge attribute
column from which values will be copied.}

\item{edge_attr_to}{the name of the new edge
attribute column to which the copied values will be
placed.}
}
\value{
a graph object of class
\code{dgr_graph}.
}
\description{
Within a graph's internal edge data
frame (edf), copy the contents an existing edge
attribute and create a distinct edge attribute
within the edf with a different attribute name.
}
\examples{
# Create a random graph
graph <-
  create_random_graph(
  5, 8, set_seed = 3) \%>\%
  set_edge_attrs("color", "green")

# Get the graph's internal edf to show which
# edge attributes are available
get_edge_df(graph)
#>   from to rel color
#> 1    4  1     green
#> 2    2  4     green
#> 3    4  3     green
#> 4    3  5     green
#> 5    5  1     green
#> 6    4  5     green
#> 7    2  1     green
#> 8    3  2     green

# Make a copy the `color` edge attribute as
# the `color_2` edge attribute
graph <-
  graph \%>\%
  copy_edge_attrs("color", "color_2")

# Get the graph's internal edf to show that the
# edge attribute had been copied
get_edge_df(graph)
#>   from to rel color color_2
#> 1    4  1     green   green
#> 2    2  4     green   green
#> 3    4  3     green   green
#> 4    3  5     green   green
#> 5    5  1     green   green
#> 6    4  5     green   green
#> 7    2  1     green   green
#> 8    3  2     green   green
}

