% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_asymmetric_nodes.R
\name{count_asymmetric_nodes}
\alias{count_asymmetric_nodes}
\title{Get the number asymmetrically-connected node pairs}
\usage{
count_asymmetric_nodes(graph)
}
\arguments{
\item{graph}{a graph object of class
\code{dgr_graph}.}
}
\value{
a single numeric value representing the
mean distance of the graph.
}
\description{
Get the number mutally-connected
node pairs. This works for directed graphs.
}
\examples{
# Create a cycle graph
graph <-
  create_graph() \%>\%
  add_cycle(n = 5)

# Get a count of asymmetrically-
# connected node pairs
count_asymmetric_nodes(graph)
#> [1] 5

# Create a full graph and then
# count the asymmetrically-
# connected node pairs
create_graph() \%>\%
  add_full_graph(n = 10) \%>\%
  count_asymmetric_nodes()
#> [1] 0
}
