% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/country_graph.R
\name{country_graph}
\alias{country_graph}
\title{Create a graph object that contains the boundaries of a country}
\usage{
country_graph(iso_a2 = NULL, option = 1)
}
\arguments{
\item{iso_a2}{the ISO 2-letter identifier for a country.}

\item{option}{an integer that is used for alternate views of certain
countries.}
}
\value{
a graph object of class \code{dgr_graph}.
}
\description{
Create a graph object that contains the boundaries of a country.
}

