% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create_graph_series.R
\name{create_graph_series}
\alias{create_graph_series}
\title{Create a graph series object
Create a graph series object for storage of multiple graphs across a sequential or temporal one-dimensional array.}
\usage{
create_graph_series(graph = NULL, graph_name = NULL, graph_time = NULL,
  graph_tz = NULL, series_name = NULL, series_type = "sequential")
}
\arguments{
\item{graph}{a graph object to add to the new graph series object}

\item{graph_name}{an optional vector of strings for labeling the graphs added to the graph series.}

\item{graph_time}{a vector of date or date-time strings which is required for a graph series of the type 'temporal'.}

\item{graph_tz}{an optional value for the time zone (tz) value corresponding to the date or date-time string supplied as values to 'graph_time'. If no time zone is set, then the times will be set to 'GMT'.}

\item{series_name}{an optional name to ascribe to the series.}

\item{series_type}{either a 'sequential' type (the default) or a 'temporal' type (which requires date-time strings and time zone codes to be supplied).}
}
\value{
a graph series object of type 'dgr_graph_1D'.
}
\description{
Create a graph series object
Create a graph series object for storage of multiple graphs across a sequential or temporal one-dimensional array.
}

