% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create_subgraph.R
\name{create_subgraph}
\alias{create_subgraph}
\title{Create a subgraph based on a walk distance from a specified node}
\usage{
create_subgraph(graph, starting_node, distance)
}
\arguments{
\item{graph}{a graph object of class \code{dgr_graph} that is created using \code{create_graph}.}

\item{starting_node}{the node from which the subgraph will originate.}

\item{distance}{the maximum number of steps from the \code{starting_node} for inclusion in the subgraph.}
}
\description{
Create a subgraph for a neighborhood of nodes connected a specified distance from the selected node.
}

