% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create_subgraph_from_selection.R
\name{create_subgraph_from_selection}
\alias{create_subgraph_from_selection}
\title{Create a subgraph based on a selection of nodes or edges}
\usage{
create_subgraph_from_selection(graph)
}
\arguments{
\item{graph}{a graph object of class \code{dgr_graph} that is created
using \code{create_graph}.}
}
\value{
a graph object of class \code{dgr_graph}.
}
\description{
Create a subgraph based on a selection of nodes or edges\
extant in the graph object.
}

