% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_xy_err.R
\name{create_xy_err}
\alias{create_xy_err}
\title{Create a series of (x, y) error values for a graph}
\usage{
create_xy_err(xy_pts, x_err = NULL, y_err = NULL, line_width = 1.5,
  line_color = "gray", x_err_pos = NULL, x_err_neg = NULL,
  y_err_pos = NULL, y_err_neg = NULL)
}
\arguments{
\item{xy_pts}{an NDF containing (x, y) for graphing}

\item{x_err}{a vector of error values for associated
x values.}

\item{y_err}{a vector of error values for associated
y values.}

\item{line_width}{the line width for the error lines.}

\item{line_color}{the color of error lines.}

\item{x_err_pos}{a vector of upper error values for
associated x values.}

\item{x_err_neg}{a vector of lower error values for
associated x values.}

\item{y_err_pos}{a vector of upper error values for
associated y values.}

\item{y_err_neg}{a vector of lower error values for
associated y values.}
}
\description{
Create a series of (x, y) error values
along for associated (x, y) data points in a series
for plotting on a graph.
}

