% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_xy_lines.R
\name{create_xy_lines}
\alias{create_xy_lines}
\title{Create a joining lines for (x, y) points}
\usage{
create_xy_lines(xy_pts, line_width = 2.5, line_color = "gray50")
}
\arguments{
\item{xy_pts}{an NDF containing (x, y) for graphing}

\item{line_width}{the line width.}

\item{line_color}{the line color.}
}
\description{
Create lines for (x, y) points, along
with other style attributes, for the purpose of
graph plotting
}

