% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_xy_pts.R
\name{create_xy_pts}
\alias{create_xy_pts}
\title{Create a series of (x, y) points for a graph}
\usage{
create_xy_pts(series_label, x, y, shape = "circle", line_width = 2.5,
  fill_color = "white", line_color = "black", width = 0.15,
  height = 0.15)
}
\arguments{
\item{series_label}{a name for the series of (x, y)
points.}

\item{x}{a vector of x points.}

\item{y}{a vector of y points.}

\item{shape}{the shape for the points.}

\item{line_width}{the line width for the shape representing the
data point.}

\item{fill_color}{the fill color for the shape representing the
data point.}

\item{line_color}{the color of the shape line representing the
data point.}

\item{width}{the width of the shape representing the
data point.}

\item{height}{the height of the shape representing the
data point.}
}
\description{
Create a series of (x, y) points along
with other style attributes for plotting on a graph
}

