% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{currencies}
\alias{currencies}
\title{ISO-4217 currency data.}
\format{A data frame with 171 rows and 4 variables:
\describe{
  \item{iso_4217_code}{the three-letter currency code according to the ISO-4217 standard}
  \item{curr_number}{the three-digit code number assigned to each currency under the ISO-4217 standard}
  \item{exponent}{the base 10 exponent of the minor currency unit in relation to the major currency unit (it can be assumed also to be number of decimal places that is commonly considered for the currency)}
  \item{currency_name}{the English name of the currency}
}}
\source{
\url{https://en.wikipedia.org/wiki/ISO_4217}
}
\usage{
currencies
}
\description{
A dataset containing currency information from the ISO-4217 standard.
}
\keyword{datasets}
