% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currency.R
\name{currency}
\alias{currency}
\title{Currency codes and currency representation attributes}
\usage{
currency()
}
\description{
Create a data frame containing attributes
of ISO 4217 currency codes and the accepted means to
represent the monetary values
}

