% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/days_in_data_range.R
\name{days_in_data_range}
\alias{days_in_data_range}
\title{Calculate the number of days in a within a vector
of dates}
\usage{
days_in_data_range(dates)
}
\arguments{
\item{dates}{a vector of \code{character} objects
in the format \code{YYYY-MM-DD} or \code{numeric}
objects where the values are integer numbers of
hours.}
}
\description{
Calculate the number of days spanning
a vector of dates when given those dates as either
\code{character} objects or as \code{numeric}
objects.
}

