% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/days_in_month.R
\name{days_in_month}
\alias{days_in_month}
\title{Calculate the number of days in a month}
\usage{
days_in_month(date)
}
\arguments{
\item{date}{either a \code{character} object in the
format \code{YYYY-MM-DD} or a \code{numeric} object
where the value is the integer number of hours.}
}
\description{
Calculate the number of days in a
month when given either a date as a \code{character}
object or as a \code{numeric} object.
}

