% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_nodes_in_selection.R
\name{delete_nodes_in_selection}
\alias{delete_nodes_in_selection}
\title{Delete all selected nodes}
\usage{
delete_nodes_in_selection(graph)
}
\arguments{
\item{graph}{a graph object of class \code{dgr_graph} that is created
using \code{create_graph}.}
}
\value{
a graph object of class \code{dgr_graph}.
}
\description{
In a graph object of class \code{dgr_graph}, delete
all nodes present in a selection.
}

