% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_nodes_ws.R
\name{delete_nodes_ws}
\alias{delete_nodes_ws}
\title{Delete all selected nodes in a node selection}
\usage{
delete_nodes_ws(graph)
}
\arguments{
\item{graph}{a graph object of class
\code{dgr_graph} that is created using
\code{create_graph}.}
}
\value{
a graph object of class \code{dgr_graph}.
}
\description{
In a graph object of class
\code{dgr_graph}, delete all nodes present in a
selection.
}
\examples{
# Create a graph
graph <-
  create_graph() \%>\%
  add_n_nodes(3) \%>\%
  add_edges_w_string(
    "1->3 1->2 2->3")

# Select node with ID `1`
graph <-
  graph \%>\%
  select_nodes_by_id(1)

# Delete node in selection (and any attached edges)
graph <-
  graph \%>\%
  delete_nodes_ws

# Get a count of nodes in the graph
node_count(graph)
#> [1] 2
}

