% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deposit_edge_count_in_selection.R
\name{deposit_edge_count_in_selection}
\alias{deposit_edge_count_in_selection}
\title{Deposit a count of edges (available in a selection) in the graph}
\usage{
deposit_edge_count_in_selection(graph)
}
\arguments{
\item{graph}{a graph object of class \code{dgr_graph} that is created
using \code{create_graph}.}
}
\value{
a graph object of class \code{dgr_graph}.
}
\description{
From a graph object of class \code{dgr_graph}, get a
count of edges available in a selection and deposit that value in
the graph for later retrieval using \code{withdraw_values}.
}

