% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deposit_node_attr_from_selection.R
\name{deposit_node_attr_from_selection}
\alias{deposit_node_attr_from_selection}
\title{Deposit node attributes (based on a selection of nodes) in the graph}
\usage{
deposit_node_attr_from_selection(graph, node_attr, mode = NULL)
}
\arguments{
\item{graph}{a graph object of class \code{dgr_graph} that is created
using \code{create_graph}.}

\item{node_attr}{the node attribute from which to obtain values.}

\item{mode}{a option to recast the returned vector of node attribute
value as \code{numeric} or \code{character}.}
}
\value{
a graph object of class \code{dgr_graph}.
}
\description{
From a graph object of class \code{dgr_graph}, get node
attribute properties for nodes available in a selection and deposit
those values in the graph for later retrieval using \code{withdraw_values}.
}

