% Generated by roxygen2 (4.0.1): do not edit by hand
\name{diagram_examples}
\alias{diagram_examples}
\title{Several examples of diagrams that can be created}
\usage{
diagram_examples(example_number)
}
\arguments{
\item{example_number}{an integer from 1 to 5 that specifies which example to display.}
}
\description{
This function provides several examples for diagramming code and output.
}

