% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiagrammeR-package.R
\docType{package}
\name{DiagrammeR-package}
\alias{DiagrammeR-package}
\title{DiagrammeR: Graph/Network Visualization}
\description{
Build graph/network structures using functions for stepwise addition and deletion of nodes and edges. Work with data available in tables for bulk addition of nodes, edges, and associated metadata. Use graph selections and traversals to apply changes to specific nodes or edges. A wide selection of graph algorithms allow for the analysis of graphs. Visualize the graphs and take advantage of any aesthetic properties assigned to nodes and edges.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rich-iannone.github.io/DiagrammeR/}
  \item \url{https://github.com/rich-iannone/DiagrammeR}
  \item Report bugs at \url{https://github.com/rich-iannone/DiagrammeR/issues}
}

}
\author{
\strong{Maintainer}: Richard Iannone \email{riannone@me.com} (\href{https://orcid.org/0000-0003-3925-190X}{ORCID})

Authors:
\itemize{
  \item Olivier Roy \email{olivierroy71@hotmail.com}
}

}
\keyword{internal}
