% Generated by roxygen2 (4.0.2): do not edit by hand
\name{DiagrammeR}
\alias{DiagrammeR}
\title{R + mermaid.js}
\usage{
DiagrammeR(diagram = "", width = NULL, height = NULL)
}
\arguments{
\item{diagram}{string diagram in mermaid markdown-like language.
If no diagram is provided \code{diagram = ""} then the function will assume that
a diagram will be provided by \code{\link[htmltools]{tags}} and
\code{DiagrammeR} is just being used for dependency injection.}
}
\description{
Make diagrams in R using \href{https://github.com/knsv/mermaid/wiki}{mermaid.js}
with infrastructure provided by \href{http://www.htmlwidgets.org/}{htmlwidgets}.
}

