% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_aes.R
\name{edge_aes}
\alias{edge_aes}
\title{Insert edge aesthetic attributes during edge creation}
\usage{
edge_aes(style = NULL, penwidth = NULL, color = NULL, arrowsize = NULL,
  arrowhead = NULL, arrowtail = NULL, fontname = NULL, fontsize = NULL,
  fontcolor = NULL, len = NULL, minlen = NULL, tooltip = NULL,
  URL = NULL, layer = NULL, label = NULL, labelfontname = NULL,
  labelfontsize = NULL, labelfontcolor = NULL, labelangle = NULL,
  labeldistance = NULL, labelfloat = NULL, labeltooltip = NULL,
  labelhref = NULL, labelURL = NULL, labeltarget = NULL,
  edgetooltip = NULL, edgehref = NULL, edgeURL = NULL,
  edgetarget = NULL, headtooltip = NULL, headhref = NULL,
  headURL = NULL, headtarget = NULL, headclip = NULL, headlabel = NULL,
  headport = NULL, tailtooltip = NULL, tailhref = NULL, tailURL = NULL,
  tailtarget = NULL, tailclip = NULL, taillabel = NULL, tailport = NULL,
  dir = NULL, target = NULL, weight = NULL, constraint = NULL,
  decorate = NULL, href = NULL, lhead = NULL, ltail = NULL,
  samehead = NULL, sametail = NULL)
}
\arguments{
\item{style}{style}

\item{penwidth}{penwidth}

\item{color}{color}

\item{arrowsize}{arrowsize}

\item{arrowhead}{arrowhead}

\item{arrowtail}{arrowtail}

\item{fontname}{fontname}

\item{fontsize}{fontsize}

\item{fontcolor}{fontcolor}

\item{len}{len}

\item{minlen}{minlen}

\item{tooltip}{tooltip}

\item{URL}{URL}

\item{layer}{layer}

\item{label}{label}

\item{labelfontname}{labelfontname}

\item{labelfontsize}{labelfontsize}

\item{labelfontcolor}{labelfontcolor}

\item{labelangle}{labelangle}

\item{labeldistance}{labeldistance}

\item{labelfloat}{labelfloat}

\item{labeltooltip}{labeltooltip}

\item{labelhref}{labelhref}

\item{labelURL}{labelURL}

\item{labeltarget}{labeltarget}

\item{edgetooltip}{edgetooltip}

\item{edgehref}{edgehref}

\item{edgeURL}{edgeURL}

\item{edgetarget}{edgetarget}

\item{headtooltip}{headtooltip}

\item{headhref}{headhref}

\item{headURL}{headURL}

\item{headtarget}{headtarget}

\item{headclip}{headclip}

\item{headlabel}{headlabel}

\item{headport}{headport}

\item{tailtooltip}{tailtooltip}

\item{tailhref}{tailhref}

\item{tailURL}{tailURL}

\item{tailtarget}{tailtarget}

\item{tailclip}{tailclip}

\item{taillabel}{taillabel}

\item{tailport}{tailport}

\item{dir}{dir}

\item{target}{target}

\item{weight}{weight}

\item{constraint}{constraint}

\item{decorate}{decorate}

\item{href}{href}

\item{lhead}{lhead}

\item{ltail}{ltail}

\item{samehead}{samehead}

\item{sametail}{sametail}
}
\description{
This helper function should be
invoked to provide values for the namesake
\code{edge_aes} argument, which is present
in any function where edges are created.
}
\examples{
# Create a new graph and add
# a path with several edge
# aesthetic attributes
graph <-
  create_graph() \%>\%
  add_path(
    n = 3,
    type = "path",
    edge_aes = edge_aes(
      style = "dot",
      color = c("red", "blue")))

# View the graph's internal
# node data frame; the node
# aesthetic attributes have
# been inserted
graph \%>\%
  get_edge_df()
#>   id from to  rel style color
#> 1  1    1  2 <NA>   dot   red
#> 2  2    2  3 <NA>   dot  blue
}
