% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_info.R
\name{edge_info}
\alias{edge_info}
\title{Get detailed information on edges}
\usage{
edge_info(graph)
}
\arguments{
\item{graph}{a graph object of class \code{dgr_graph}.}
}
\value{
a data frame containing information specific to each edge within
the graph.
}
\description{
Obtain a data frame with detailed information on edges and
their interrelationships within a graph.
}
\examples{
# Set a seed
set.seed(24)

# Create a node data frame (ndf)
nodes <-
  create_node_df(
    n = 26,
    label = TRUE,
    type = c(rep("a", 7),
             rep("b", 9),
             rep("c", 8),
             rep("d", 2)))

# Create an edge data frame (edf)
edges <-
  create_edge_df(
    from = sample(1:26, replace = TRUE),
    to = sample(1:26, replace = TRUE),
    rel = c(rep("rel_a", 7),
            rep("rel_b", 9),
            rep("rel_c", 8),
            rep("rel_d", 2)))

# Create a graph using the ndf and edf
graph <-
  create_graph(
    nodes_df = nodes,
    edges_df = edges)

# Get a data frame with information about
# the graph's edges
edge_info(graph)
#>    from to   rel
#> 1     8  2 rel_a
#> 2     6 16 rel_a
#> 3    19 17 rel_a
#> 4    14  2 rel_a
#> 5    18  9 rel_a
#>..   ... ..   ...
}

