% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{edge_list_1}
\alias{edge_list_1}
\title{Edge list - Version 1.}
\format{A data frame with 19 rows and 2 variables:
\describe{
  \item{from}{integer values that state the node ID values where an edge starts}
  \item{to}{integer values that state the node ID values where an edge terminates}
}}
\usage{
edge_list_1
}
\description{
A very simple, 2-column data frame that can be used to generate graph edges.
}
\keyword{datasets}
