% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/edge_relationship.R
\name{edge_relationship}
\alias{edge_relationship}
\title{Create, read, update, delete, or report status of an edge relationship
From a graph object of class 'gv_graph', query an edge in the graph (defined by a pair of node IDs extant in the graph) and perform operations on the relationship for that edge.}
\usage{
edge_relationship(graph, from, to, action = "read", value = NULL)
}
\arguments{
\item{graph}{a graph object of class 'gv_graph' that is created using 'create_graph'.}

\item{from}{a node ID from which the edge to be queried is outgoing.}

\item{to}{a node ID to which the edge to be queried is incoming.}

\item{action}{the type of operation to perform post-query. To remove a relationship from an edge, use either 'delete', 'remove', or 'drop'. To add a relationship to an edge with no set relationship, use 'add' or 'create'. To update an edge relationship, use 'update'. To return the value of an edge relationship, use 'read'. To determine whether there is a set relationship, use 'check'.}

\item{value}{a string denoting the relationship, to be supplied when either adding or updating an edge relationship.}
}
\value{
a graph object of class 'gv_graph'.
}
\description{
Create, read, update, delete, or report status of an edge relationship
From a graph object of class 'gv_graph', query an edge in the graph (defined by a pair of node IDs extant in the graph) and perform operations on the relationship for that edge.
}

