% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_dot.R
\name{generate_dot}
\alias{generate_dot}
\title{Generate DOT code using a graph object}
\usage{
generate_dot(graph)
}
\arguments{
\item{graph}{a graph object of class \code{dgr_graph}.}
}
\value{
a character vector of length 1 containing Graphviz DOT code.
}
\description{
Generates Graphviz DOT code as an R character object using DiagrammeR graph
object.
}
