% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_betweenness.R
\name{get_betweenness}
\alias{get_betweenness}
\title{Get betweenness centrality scores}
\usage{
get_betweenness(graph)
}
\arguments{
\item{graph}{a graph object of class
\code{dgr_graph}.}
}
\value{
a data frame with betweenness scores for
each of the nodes.
}
\description{
Get the betweenness centrality scores
for all nodes in a graph.
}
\examples{
# Create a random graph using the
# `add_gnm_graph()` function
graph <-
  create_graph() \%>\%
  add_gnm_graph(
    n = 10,
    m = 12,
    set_seed = 23)

# Get the betweenness scores
# for nodes in the graph
get_betweenness(graph)
#>    id betweenness
#> 1   1           0
#> 2   2           7
#> 3   3           0
#> 4   4           0
#> 5   5           5
#> 6   6           0
#> 7   7           0
#> 8   8           1
#> 9   9           0
#> 10 10           0

# Add the betweenness
# values to the graph
# as a node attribute
graph <-
  graph \%>\%
  join_node_attrs(
    df = get_betweenness(.))

# Display the graph's node
# data frame
get_node_df(graph)
#>    id type label betweenness
#> 1   1 <NA>  <NA>           0
#> 2   2 <NA>  <NA>           7
#> 3   3 <NA>  <NA>           0
#> 4   4 <NA>  <NA>           0
#> 5   5 <NA>  <NA>           5
#> 6   6 <NA>  <NA>           0
#> 7   7 <NA>  <NA>           0
#> 8   8 <NA>  <NA>           1
#> 9   9 <NA>  <NA>           0
#> 10 10 <NA>  <NA>           0
}
