% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bridging.R
\name{get_bridging}
\alias{get_bridging}
\title{Get bridging scores}
\usage{
get_bridging(graph)
}
\arguments{
\item{graph}{a graph object of class
\code{dgr_graph}.}
}
\value{
a data frame with bridging scores for
each of the nodes.
}
\description{
Get the bridging scores (based on
Valente's Bridging vertex measure) for all nodes in
a graph.
}
\examples{
\dontrun{
# Create a random graph
graph <-
  create_random_graph(
    10, 22, set_seed = 1)

# Get the bridging scores for nodes in the graph
get_bridging(graph)
#>    node bridging
#> 1     1    6.633333
#> 2     2    5.638095
#> 3     3    1.904762
#> 4     4    4.019048
#> 5     5    8.157143
#> 6     6    2.000000
#> 7     7   10.157143
#> 8     8    8.857143
#> 9     9    3.466667
#> 10   10    1.166667
}
}

