% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cache.R
\name{get_cache}
\alias{get_cache}
\title{Get a cached vector from a graph object}
\usage{
get_cache(graph)
}
\arguments{
\item{graph}{a graph object of class
\code{dgr_graph} that is created using
\code{create_graph}.}
}
\value{
a vector.
}
\description{
Get the vector cached in a graph object
of class \code{dgr_graph}.
}

