% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cmty_edge_btwns.R
\name{get_cmty_edge_btwns}
\alias{get_cmty_edge_btwns}
\title{Get community membership by edge betweenness}
\usage{
get_cmty_edge_btwns(graph)
}
\arguments{
\item{graph}{a graph object of class
\code{dgr_graph}.}
}
\value{
a data frame with group membership
assignments for each of the nodes.
}
\description{
Using edge betweenness, obtain the
group membership values for each of the nodes in
the graph.
}
\examples{
# Create a random graph using the
# `add_gnm_graph()` function
graph <-
  create_graph(
    directed = FALSE) \%>\%
  add_gnm_graph(
    n = 10,
    m = 15,
    set_seed = 23)

# Get the group membership values for all nodes
# in the graph through calculation of the leading
# non-negative eigenvector of the modularity matrix
# of the graph
get_cmty_edge_btwns(graph)
#>    id edge_btwns_group
#> 1   1                1
#> 2   2                2
#> 3   3                2
#> 4   4                3
#> 5   5                2
#> 6   6                2
#> 7   7                1
#> 8   8                3
#> 9   9                3
#> 10 10                4

# Add the group membership
# values to the graph
# as a node attribute
graph <-
  graph \%>\%
  join_node_attrs(
     df = get_cmty_edge_btwns(.))

# Display the graph's node data frame
get_node_df(graph)
#>    id type label edge_btwns_group
#> 1   1 <NA>  <NA>                1
#> 2   2 <NA>  <NA>                2
#> 3   3 <NA>  <NA>                2
#> 4   4 <NA>  <NA>                3
#> 5   5 <NA>  <NA>                2
#> 6   6 <NA>  <NA>                2
#> 7   7 <NA>  <NA>                1
#> 8   8 <NA>  <NA>                3
#> 9   9 <NA>  <NA>                3
#> 10 10 <NA>  <NA>                4
}
