% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cmty_walktrap.R
\name{get_cmty_walktrap}
\alias{get_cmty_walktrap}
\title{Get community membership using the Walktrap method}
\usage{
get_cmty_walktrap(graph, steps = 4)
}
\arguments{
\item{graph}{a graph object of class
\code{dgr_graph}.}

\item{steps}{the number of steps to take for each
of the random walks.}
}
\value{
a data frame with group membership
assignments for each of the nodes.
}
\description{
With the Walktrap community finding
algorithm, obtain the group membership values for
each of the nodes in the graph.
}
\examples{
# Create a random graph using the
# `add_gnm_graph()` function
graph <-
  create_graph(
    directed = FALSE) \%>\%
  add_gnm_graph(
    n = 10,
    m = 15,
    set_seed = 23)

# Get the group membership
# values for all nodes in the
# graph through the Walktrap
# community finding algorithm
get_cmty_walktrap(graph)
#>    id walktrap_group
#> 1   1              3
#> 2   2              2
#> 3   3              2
#> 4   4              1
#> 5   5              2
#> 6   6              2
#> 7   7              3
#> 8   8              1
#> 9   9              1
#> 10 10              4

# Add the group membership
# values to the graph as a
# node attribute
graph <-
  graph \%>\%
  join_node_attrs(
    df = get_cmty_walktrap(.))

# Display the graph's node data frame
get_node_df(graph)
#>    id type label walktrap_group
#> 1   1 <NA>  <NA>              3
#> 2   2 <NA>  <NA>              2
#> 3   3 <NA>  <NA>              2
#> 4   4 <NA>  <NA>              1
#> 5   5 <NA>  <NA>              2
#> 6   6 <NA>  <NA>              2
#> 7   7 <NA>  <NA>              3
#> 8   8 <NA>  <NA>              1
#> 9   9 <NA>  <NA>              1
#> 10 10 <NA>  <NA>              4
}
