% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coreness.R
\name{get_coreness}
\alias{get_coreness}
\title{Get coreness values for graph nodes}
\usage{
get_coreness(graph, direction = "all")
}
\arguments{
\item{graph}{a graph object of class
\code{dgr_graph}.}

\item{direction}{using \code{all} (the default), the
search will ignore edge direction while traversing
through the graph. With \code{out}, measurements of
paths will be from a node whereas with \code{in},
measurements of paths will be to a node.}
}
\value{
a data frame with coreness values for
each of the nodes.
}
\description{
Get the coreness values for all nodes
in a graph.
}
\examples{
# Create a random graph using the
# `add_gnm_graph()` function
graph <-
  create_graph(
    directed = FALSE) \%>\%
  add_gnm_graph(
    n = 10,
    m = 15,
    set_seed = 23)

# Get coreness values for
# all nodes in the graph
get_coreness(graph)
#>    id coreness
#> 1   1        3
#> 2   2        3
#> 3   3        2
#> 4   4        3
#> 5   5        3
#> 6   6        3
#> 7   7        3
#> 8   8        2
#> 9   9        2
#> 10 10        0

# Add the coreness values
# to the graph as a node
# attribute
graph <-
  graph \%>\%
  join_node_attrs(
    df = get_coreness(.))

# Display the graph's node data frame
get_node_df(graph)
#>    id type label coreness
#> 1   1 <NA>  <NA>        3
#> 2   2 <NA>  <NA>        3
#> 3   3 <NA>  <NA>        2
#> 4   4 <NA>  <NA>        3
#> 5   5 <NA>  <NA>        3
#> 6   6 <NA>  <NA>        3
#> 7   7 <NA>  <NA>        3
#> 8   8 <NA>  <NA>        2
#> 9   9 <NA>  <NA>        2
#> 10 10 <NA>  <NA>        0
}
