% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_edge_attr_from_selection.R
\name{get_edge_attr_from_selection}
\alias{get_edge_attr_from_selection}
\title{Get edge attributes based on a selection of edges}
\usage{
get_edge_attr_from_selection(graph)
}
\arguments{
\item{graph}{a graph object of class \code{dgr_graph} that is created
using \code{create_graph}.}
}
\value{
a graph object of class \code{dgr_graph}.
}
\description{
From a graph object of class \code{dgr_graph}, get edge
attribute properties for edges available in a selection.
}

