% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_edge_attrs_ws.R
\name{get_edge_attrs_ws}
\alias{get_edge_attrs_ws}
\title{Get edge attribute values}
\usage{
get_edge_attrs_ws(graph, edge_attr)
}
\arguments{
\item{graph}{a graph object of class
\code{dgr_graph}.}

\item{edge_attr}{the name of the attribute for which
to get values.}

\item{from}{an optional vector of node IDs from
which the edge is outgoing for filtering the list
of edges.}

\item{to}{an optional vector of node IDs from which
the edge is incoming for filtering the list of
edges.}
}
\value{
a named vector of edge attribute values for
the attribute given by \code{edge_attr} by edge.
}
\description{
From a graph object of class
\code{dgr_graph}, get edge attribute values for one
or more edges.
}
\examples{
# Create a simple graph where edges have an edge
# attribute named `value`
graph <-
  create_graph() \%>\%
  add_n_nodes(4) \%>\%
  {
    edges <-
      create_edge_df(
        from = c(1, 2, 1, 4),
        to = c(2, 3, 4, 3),
        rel = "rel")
    add_edge_df(., edges)
  } \%>\%
  set_edge_attrs(
    "value", 1.6, 1, 2) \%>\%
  set_edge_attrs(
    "value", 4.3, 1, 4) \%>\%
  set_edge_attrs(
    "value", 2.9, 2, 3) \%>\%
  set_edge_attrs(
    "value", 8.4, 4, 3)

# Select the edges defined as `1`->`3`
# and `2`->`3`
graph <-
  graph \%>\%
  select_edges(
    from = c(1, 2),
    to = c(2, 3))

# To only return edge attribute values for specified
# edges, use the `from` and `to` arguments
graph \%>\%
  get_edge_attrs_ws(
    edge_attr = "value")
#> 1->2 2->3
#>  1.6  2.9
}
