% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_global_graph_attr.R
\name{get_global_graph_attr}
\alias{get_global_graph_attr}
\title{Get global graph attributes}
\usage{
get_global_graph_attr(graph)
}
\arguments{
\item{graph}{a graph object of class \code{dgr_graph} that is created using
\code{create_graph}.}
}
\value{
a list object containing global attributes for a graph.
}
\description{
Get the presently set global attributes for a graph object of
class \code{dgr_graph}.
}
\examples{
\dontrun{
}
}

