% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_graph_actions.R
\name{get_graph_actions}
\alias{get_graph_actions}
\title{Get information on any available graph actions}
\usage{
get_graph_actions(graph)
}
\arguments{
\item{graph}{a graph object of class
\code{dgr_graph}.}
}
\value{
a \code{df_tbl} object.
}
\description{
Get a tibble of the available graph
actions, which contains information on function
invocations to be called on the graph at every
transformation step, or, when manually invoked
with the \code{trigger_graph_actions()} function.
}
