% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_graph_log.R
\name{get_graph_log}
\alias{get_graph_log}
\title{Get the graph log information}
\usage{
get_graph_log(graph)
}
\arguments{
\item{graph}{a graph object of class
\code{dgr_graph}.}
}
\value{
a \code{df_tbl} object.
}
\description{
Get a tibble of the graph log, which
contains information on the functions called on
the graph that resulted in some transformation of
the graph.
}
