% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_jaccard_similarity.R
\name{get_jaccard_similarity}
\alias{get_jaccard_similarity}
\title{Get Jaccard similarity coefficient scores}
\usage{
get_jaccard_similarity(graph, nodes = NULL, direction = "all")
}
\arguments{
\item{graph}{a graph object of class
\code{dgr_graph}.}

\item{nodes}{an optional vector of node IDs to
consider for Jaccard similarity scores. If not
supplied, then similarity scores will be provided
for every pair of nodes in the graph.}

\item{direction}{using \code{all} (the default), the
function will ignore edge direction when
determining scores for neighboring nodes. With
\code{out} and \code{in}, edge direction for
neighboring nodes will be considered.}
}
\value{
a data frame with Jaccard similiarity values
for each pair of nodes considered.
}
\description{
Get the Jaccard similiarity coefficient
scores for one or more nodes in a graph.
}
\examples{
library(magrittr)

# Create a random graph
graph <-
  create_random_graph(
    10, 22, set_seed = 1)

# Get the Jaccard similarity values for
# nodes `5`, `6`, and `7`
get_jaccard_similarity(graph, 5:7)
#>   node       n_5       n_6       n_7
#> 1    5 1.0000000 0.2857143 0.5000000
#> 2    6 0.2857143 1.0000000 0.2857143
#> 3    7 0.5000000 0.2857143 1.0000000
}

