% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_min_spanning_tree.R
\name{get_min_spanning_tree}
\alias{get_min_spanning_tree}
\title{Get a minimum spanning tree subgraph}
\usage{
get_min_spanning_tree(graph)
}
\arguments{
\item{graph}{a graph object of class
\code{dgr_graph}.}
}
\value{
a graph object of class \code{dgr_graph}.
}
\description{
Get a minimum spanning tree subgraph
for a connected graph of class \code{dgr_graph}.
}

