% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_node_attr.R
\name{get_node_attr}
\alias{get_node_attr}
\title{Get node attributes}
\usage{
get_node_attr(graph, node_attr, mode = NULL, nodes = NULL)
}
\arguments{
\item{graph}{a graph object of class \code{dgr_graph}.}

\item{node_attr}{the node attribute from which to obtain values.}

\item{mode}{a option to recast the returned vector of node attribute
value as \code{numeric} or \code{character}.}

\item{nodes}{an optional vector of node IDs for filtering list of
nodes present in the graph.}
}
\value{
a graph object of class \code{dgr_graph}.
}
\description{
From a graph object of class \code{dgr_graph}, get node
attribute properties for one or more nodes.
}
\examples{
\dontrun{
# Create a simple graph
nodes <-
  create_nodes(nodes = c("a", "b", "c", "d"),
               type = "letter",
               label = TRUE,
               value = c(3.5, 2.6, 9.4, 2.7))

edges <-
  create_edges(from = c("a", "b", "c"),
               to = c("d", "c", "a"),
               rel = "leading_to",
               color = c("pink", "blue", "red"))

graph <-
  create_graph(nodes_df = nodes,
               edges_df = edges)

# Get node attributes for all nodes in the graph
get_node_attr(x = graph)

# Get node attributes for nodes "a" and "c" in the graph
get_node_attr(x = graph, nodes = c("a", "c"))
}
}

