% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_node_attr_from_selection.R
\name{get_node_attr_from_selection}
\alias{get_node_attr_from_selection}
\title{Get node attributes based on a selection of nodes}
\usage{
get_node_attr_from_selection(graph)
}
\arguments{
\item{graph}{a graph object of class \code{dgr_graph} that is created
using \code{create_graph}.}
}
\value{
a graph object of class \code{dgr_graph}.
}
\description{
From a graph object of class \code{dgr_graph}, get node
attribute properties for nodes available in a selection.
}

