% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_node_count_in_selection.R
\name{get_node_count_in_selection}
\alias{get_node_count_in_selection}
\title{Get a count of nodes available in a selection}
\usage{
get_node_count_in_selection(graph)
}
\arguments{
\item{graph}{a graph object of class \code{dgr_graph} that is created
using \code{create_graph}.}
}
\value{
a graph object of class \code{dgr_graph}.
}
\description{
From a graph object of class \code{dgr_graph}, get a
count of nodes available in a selection.
}

