% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_node_df.R
\name{get_node_df}
\alias{get_node_df}
\title{Get a node data frame from a graph}
\usage{
get_node_df(graph)
}
\arguments{
\item{graph}{a graph object of class \code{dgr_graph} that is created
using \code{create_graph}.}
}
\value{
a node data frame.
}
\description{
From a graph, obtain a node data frame with all current
node attributes.
}

