% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_nodes.R
\name{get_nodes}
\alias{get_nodes}
\title{Get node ID information from node data frames}
\usage{
get_nodes(...)
}
\arguments{
\item{nodes_df}{a data frame containing, at minimum, a column that contains node IDs for the graph. Optionally, additional columns (named as Graphviz node attributes) can be included with values for the named node attribute. These data frames can be conveniently generated using the 'create_nodes' function.}
}
\description{
Provides information on the node IDs from one or several node data frames.
}

