% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_paths.R
\name{get_paths}
\alias{get_paths}
\title{Get paths from a specified node in a directed graph}
\usage{
get_paths(graph, from)
}
\arguments{
\item{graph}{a graph object of class \code{dgr_graph}.}

\item{from}{the node from which all paths will be determined.}
}
\value{
a list of paths comprising vectors of node IDs in sequence of
traversal through the graph
}
\description{
Obtain a list of all possible paths from a given node within
a directed graph
}

