% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_periphery.R
\name{get_periphery}
\alias{get_periphery}
\title{Get nodes that form the graph periphery}
\usage{
get_periphery(graph)
}
\arguments{
\item{graph}{a graph object of class
\code{dgr_graph}.}
}
\value{
a vector of node IDs.
}
\description{
Get those nodes that are part of the
graph periphery (i.e., have the maximum eccentricity
in the graph).
}
\examples{
\dontrun{
# Get the nodes that are in the graph periphery of
# a randomly-created graph
get_periphery(
  create_random_graph(
    15, 24,
    fully_connected = TRUE,
    set_seed = 20))
#> [1] 13
}
}

