% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_radiality.R
\name{get_radiality}
\alias{get_radiality}
\title{Get radiality centrality scores}
\usage{
get_radiality(graph, direction = "all")
}
\arguments{
\item{graph}{a graph object of class
\code{dgr_graph}.}

\item{direction}{using \code{all} (the default), the
search will ignore edge direction while traversing
through the graph. With \code{out}, measurements of
paths will be from a node whereas with \code{in},
measurements of paths will be to a node.}
}
\value{
a data frame with radiality centrality
scores for each of the nodes.
}
\description{
Get the radiality centrality
for all nodes in a graph. These scores describe
the ease to which nodes can reach other nodes.
}
\examples{
# Create a random graph using the
# `add_gnm_graph()` function
graph <-
  create_graph() \%>\%
  add_gnm_graph(
    n = 10,
    m = 15,
    set_seed = 23)

# Get the radiality scores for nodes in the graph
get_radiality(graph)
#>    id radiality
#> 1   1    2.3333
#> 2   2    3.0000
#> 3   3    2.6667
#> 4   4    2.8889
#> 5   5    2.5556
#> 6   6    2.4444
#> 7   7    2.6667
#> 8   8    2.7778
#> 9   9    2.1111
#> 10 10    2.3333

# Add the radiality values
# to the graph as a node
# attribute
graph <-
  graph \%>\%
  join_node_attrs(
    df = get_radiality(.))

# Display the graph's node data frame
get_node_df(graph)
#>    id type label radiality
#> 1   1 <NA>  <NA>    2.3333
#> 2   2 <NA>  <NA>    3.0000
#> 3   3 <NA>  <NA>    2.6667
#> 4   4 <NA>  <NA>    2.8889
#> 5   5 <NA>  <NA>    2.5556
#> 6   6 <NA>  <NA>    2.4444
#> 7   7 <NA>  <NA>    2.6667
#> 8   8 <NA>  <NA>    2.7778
#> 9   9 <NA>  <NA>    2.1111
#> 10 10 <NA>  <NA>    2.3333
}
