% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_selection.R
\name{get_selection}
\alias{get_selection}
\title{Get the current selection available in a graph object}
\usage{
get_selection(graph)
}
\arguments{
\item{graph}{a graph object of class \code{dgr_graph} that is created
using \code{create_graph}.}
}
\value{
a list object with the current selection of nodes or edges.
}
\description{
Get the current selection of nodes or edges from a graph
object of class \code{dgr_graph}.
}
\examples{
\dontrun{
}
}

