% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/graphviz_graph.R
\name{graphviz_graph}
\alias{graphviz_graph}
\title{Create DOT code from two data frames and optionally render graph}
\usage{
graphviz_graph(nodes_df = NULL, edges_df = NULL, graph_attrs = NULL,
  node_attrs = NULL, edge_attrs = NULL, directed = TRUE,
  create_graph = TRUE, return_code = NULL, width = NULL, height = NULL)
}
\arguments{
\item{nodes_df}{a data frame containing, at minimum, a column (called 'nodes') which contains node IDs for the graph. Optionally, additional columns (named as Graphviz node attributes) can be included with values for the named node attribute.}

\item{edges_df}{a data frame containing, at minimum, a column (called 'edge_op') with edge operations as character strings (in the form of '[node_id] -> [node_id]). Alternatively, there may be two columns (called 'edge_from' and 'edge_to') where node IDs are provided. Optionally, additional columns (named as Graphviz edge attributes) can be included with values for the named edge attribute.}

\item{graph_attrs}{an optional vector of graph attribute statements that can serve as defaults.}

\item{node_attrs}{an optional vector of node attribute statements that can serve as defaults.}

\item{edge_attrs}{an optional vector of edge attribute statements that can serve as defaults.}

\item{directed}{with TRUE (the default) or FALSE, either directed or undirected edge operations will be generated, respectively.}

\item{create_graph}{with TRUE (the default) the function render the graph using the 'grViz' function; with FALSE, the generated Graphviz DOT code is returned as a vector object (useful for substituting in a separate 'grViz' statement).}

\item{return_code}{if not NULL, the value 'SVG' returns string of SVG XML text whereas the value 'DOT' returns valid Graphviz DOT code.}

\item{width}{an optional parameter for specifying the width of the resulting graphic in pixels.}

\item{height}{an optional parameter for specifying the height of the resulting graphic in pixels.}
}
\description{
Generates and returns either Graphviz DOT code or a Graphviz graph from two data frames:
one with nodes and the other with edges
}

