% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/graphviz_nodes_edges_df.R
\name{graphviz_nodes_edges_df}
\alias{graphviz_nodes_edges_df}
\title{Create DOT code from two data frames}
\usage{
graphviz_nodes_edges_df(nodes_df, edges_df)
}
\arguments{
\item{nodes_df}{}

\item{edges_df}{}
}
\description{
A function to generate DOT code from two data frames: one with nodes and the
other with edge operations.
}

