% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/graphviz_render.R
\name{graphviz_render}
\alias{graphviz_render}
\title{Render the graph or output in various formats}
\usage{
graphviz_render(graph, output = "graph", width = NULL, height = NULL)
}
\arguments{
\item{graph}{a 'gv_graph' object, created using the \code{graphviz_graph} function}

\item{output}{a string specifying the output type; \code{graph} (the default) renders the graph using the \code{grViz} function, \code{DOT} outputs DOT code for the graph, and \code{SVG} provides SVG code for the rendered graph.}

\item{width}{an optional parameter for specifying the width of the resulting graphic in pixels.}

\item{height}{an optional parameter for specifying the height of the resulting graphic in pixels.}
}
\description{
Using a 'gv_graph' object, either render graph in the Viewer or output in various formats.
}

