% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/graphviz_single_df.R
\name{graphviz_single_df}
\alias{graphviz_single_df}
\title{Create DOT code from a data frame
A function to generate DOT code from a single data frame}
\usage{
graphviz_single_df(df, edge_between, node_attr = NULL, edge_attr = NULL,
  add_labels = FALSE)
}
\arguments{
\item{df}{the data frame object from which node and edge statements in DOT notation are to be generated.}

\item{edge_between}{a vector object containing statements that provide information on the relationships between nodes in different columns. The basic syntax takes the form of: "df_column_name_1 [->|--] df_column_name_2".}

\item{node_attr}{a vector object containing statements for node attributes.}

\item{edge_attr}{a vector object containing statements for edge attributes.}

\item{add_labels}{whether to automatically generate a set of node and edge labels based on the node ID and the edge operation, respectively.}
}
\description{
Create DOT code from a data frame
A function to generate DOT code from a single data frame
}

