% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grViz.R
\name{grVizOutput}
\alias{grVizOutput}
\title{Widget output function for use in Shiny}
\usage{
grVizOutput(outputId, width = "100\%", height = "400px")
}
\arguments{
\item{outputId}{Output variable to read from.}

\item{width}{A valid CSS unit for the width or a number, which will be
coerced to a string and have \code{px} appended.}

\item{height}{A valid CSS unit for the height or a number, which will be
coerced to a string and have \code{px} appended.}
}
\description{
Widget output function for use in Shiny
}
