% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hex_palette.R
\name{hex_palette}
\alias{hex_palette}
\title{Create a simple color palette}
\usage{
hex_palette(number, hue_range = c(0, 360), chroma_range = c(0, 3),
  lightness_range = c(0.75, 1.5), alpha = NULL, display_colors = TRUE)
}
\arguments{
\item{hue_range}{a vector representing a range of hues from 0 to 360.}

\item{chroma_range}{a vector representing a range of chroma values from 0 to 3.}

\item{lightness_range}{a vector representing a range of lightness values from 0 to 2.}

\item{alpha}{an optional alpha value in the range of 0-100 to append to the hexadecimal color values.}
}
\value{
an edge data frame
}
\description{
Generate a randomized set of colors from the CIE Lab Color space using a set of HCL range values.
}

