% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_icon.R
\name{image_icon}
\alias{image_icon}
\title{Icons and their download locations}
\usage{
image_icon(icon_name)
}
\arguments{
\item{icon_name}{the name of the FontAwesome icon.}
}
\description{
Create a data frame with image icons and their web addresses.
}

