% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invert_selection.R
\name{invert_selection}
\alias{invert_selection}
\title{Invert selection of nodes or edges in a graph}
\usage{
invert_selection(graph)
}
\arguments{
\item{graph}{a graph object of class \code{dgr_graph} that is created
using \code{create_graph}.}
}
\value{
a graph object of class \code{dgr_graph}.
}
\description{
Modify the selection of nodes or edges within a graph
object such that all nodes or edges previously unselected will now be
selected and vice versa.
}

