% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_edge_mutual.R
\name{is_edge_mutual}
\alias{is_edge_mutual}
\title{Is the edge mutual with another edge?}
\usage{
is_edge_mutual(graph, edge)
}
\arguments{
\item{graph}{a graph object of class
\code{dgr_graph}.}

\item{edge}{an edge ID value.}
}
\value{
a logical value.
}
\description{
Determines whether an edge
definition has a mutual analogue with the
same node pair.
}
