% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_edge_attrs.R
\name{join_edge_attrs}
\alias{join_edge_attrs}
\title{Join new edge attribute values using a data frame}
\usage{
join_edge_attrs(graph, df)
}
\arguments{
\item{graph}{a graph object of class}

\item{df}{the data frame to use for joining.
\code{dgr_graph} that is created using
\code{create_graph}.}
}
\value{
a graph object of class
\code{dgr_graph}.
}
\description{
Join new edge attribute values in a
left join using a data frame. The data frame to join
should have at least one column with a name
identical to a column in the graph's edge data frame
(e.g., \code{rel} in both to join on the edge
\code{rel} values). The use of a left join in this
function means that there is no possibility that
edges in the graph might be removed after the join.
}

