% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_node_attrs.R
\name{join_node_attrs}
\alias{join_node_attrs}
\title{Join new node attribute values using a data frame}
\usage{
join_node_attrs(graph, df)
}
\arguments{
\item{graph}{a graph object of class}

\item{df}{the data frame to use for joining.
\code{dgr_graph} that is created using
\code{create_graph}.}
}
\value{
a graph object of class
\code{dgr_graph}.
}
\description{
Join new node attribute values in a
left join using a data frame. The data frame to join
should have at least one column with a name
identical to a column in the graph's node data frame
(e.g., \code{nodes} in both to join on the node ID
values). The use of a left join in this function
means that there is no possibility that nodes in the
graph might be removed after the join.
}

