% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_cond.R
\name{mk_cond}
\alias{mk_cond}
\title{Helper for making conditions for some `select_` function}
\usage{
mk_cond(...)
}
\arguments{
\item{...}{sets of 3 elements for each conditions.}
}
\value{
a string to be used for any \code{conditions}
argument.
}
\description{
Create one or multiple conditions for
certain \code{select_...} functions that have a
\code{conditions} argument.
}

