% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_aes.R
\name{node_aes}
\alias{node_aes}
\title{Insert node aesthetic attributes during node creation}
\usage{
node_aes(shape = NULL, style = NULL, penwidth = NULL, color = NULL,
  fillcolor = NULL, fontname = NULL, fontsize = NULL, fontcolor = NULL,
  peripheries = NULL, height = NULL, width = NULL, x = NULL, y = NULL,
  group = NULL, tooltip = NULL, URL = NULL, sides = NULL,
  orientation = NULL, skew = NULL, distortion = NULL,
  gradientangle = NULL, fixedsize = NULL, labelloc = NULL,
  margin = NULL)
}
\arguments{
\item{shape}{the shape to use for the node.
Some possible \code{shape} types include:
\code{circle}, \code{rectangle},
\code{triangle}, \code{plaintext},
\code{square}, and \code{polygon}.}

\item{style}{the node line style. The
\code{style} types that can be used are
\code{filled}, \code{invisible},
\code{diagonals}, \code{rounded}, \code{dashed},
\code{dotted}, \code{solid}, and \code{bold}.}

\item{penwidth}{the thickness of the stroke
line (in pt units) for the node shape. The
default value is \code{1.0}.}

\item{color}{the color of the node's
outline. Can be an X11 color or a hexadecimal
color code.}

\item{fillcolor}{the color with which
to fill the shape of the node. Can be an X11
color or a hexadecimal color code.}

\item{fontname}{the name of the
system font that will be used for any node
text.}

\item{fontsize}{the point size of
the font used for any node text.}

\item{fontcolor}{the color used
for any node text. Can be an X11 color or a
hexadecimal color code.}

\item{peripheries}{the repeated number
of node shapes (of increasing size) to
draw at the node perhipery.}

\item{height}{the height of the shape.}

\item{width}{the width of the shape.}

\item{x}{the fixed position of the node in
the x direction. Any integer-based or
floating point value will be accepted.}

\item{y}{the fixed position of the node in
the y direction. Any integer-based or
floating point value will be accepted.}

\item{group}{the node group.}

\item{tooltip}{text for a node tooltip.}

\item{URL}{a URL to associate with a node.
Upon rendering the plot, clicking nodes
with any associated URLs will open the
URL in the default browser.}

\item{sides}{when using the shape
\code{polygon}, this value will provide
the number of sides for that polygon.}

\item{orientation}{this is the angle,
in degrees, that is used to rotate
nodes that have a \code{shape} of
\code{polygon}. Not that for any of
the polygon shapes (set by the
\code{sides} node attribute), a value
for \code{orientation} that is \code{0}
results in a flat base.}

\item{skew}{a \code{0-1} value that will
result in the node shape being skewed
to the right (from bottom to top). A
value in the range \code{0} to \code{-1}
will skew the shape to the left.}

\item{distortion}{a distortion factor
that is used only when a \code{shape} of
\code{polygon} is used. A \code{0-1} value
will increasingly result in the top part
of the node polygon shape to be larger
than the bottom. Moving from \code{0}
toward \code{-1} will result in the
opposite distortion effect.}

\item{gradientangle}{the path angle for
the node color fill gradient.}

\item{fixedsize}{if set to \code{FALSE},
the size of a node is determined by
smallest width and height needed to
contain its label, if any, with a
margin specified by the \code{margin}
node attribute. The width and height
must also be at least as large as the
sizes specified by the \code{width} and
\code{height} node attributes, which
specify the minimum values. If set to
\code{TRUE}, the node size is entirely
specified by the values of the
\code{width} and \code{height} node
attributes (i.e., the node is not
expanded in size to contain the text
label).}

\item{labelloc}{sets the vertical
placement of labels for nodes and
clusters. This attribute is used only
when the height of the node is
larger than the height of its label.
The \code{labelloc} node attribute can
be set to either \code{t} (top),
\code{c} (center), or \code{b} (bottom).
By default, the label is vertically
centered.}

\item{margin}{margin}
}
\description{
This helper function should be
invoked to provide values for the namesake
\code{node_aes} argument, which is present
in any function where nodes are created.
}
\examples{
# Create a new graph and add
# a path with several node
# aesthetic attributes
graph <-
  create_graph() \%>\%
  add_path(
    n = 3,
    type = "path",
    node_aes = node_aes(
      shape = "circle",
      x = c(1, 3, 2),
      y = c(4, -1, 3)))

# View the graph's internal
# node data frame; the node
# aesthetic attributes have
# been inserted
graph \%>\%
  get_node_df()
#>   id type label  shape x  y
#> 1  1 path     1 circle 1  4
#> 2  2 path     2 circle 3 -1
#> 3  3 path     3 circle 2  3
}
