% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs_node_edge_aes_data.R
\name{node_edge_aes_data}
\alias{node_edge_aes_data}
\title{Options for node and edge aesthetics and data}
\arguments{
\item{node_aes}{an optional list of named vectors comprising node aesthetic
attributes. The helper function \code{\link{node_aes}()} is strongly
recommended for use here as it contains arguments for each of the accepted
node aesthetic attributes (e.g., \code{shape}, \code{style}, \code{color},
\code{fillcolor}).}

\item{edge_aes}{an optional list of named vectors comprising edge aesthetic
attributes. The helper function \code{\link{edge_aes}()} is strongly
recommended for use here as it contains arguments for each of the accepted
edge aesthetic attributes (e.g., \code{shape}, \code{style},
\code{penwidth}, \code{color}).}

\item{node_data}{an optional list of named vectors comprising node data
attributes. The helper function \code{\link{node_data}()} is strongly
recommended for use here as it helps bind data specifically to the created
nodes.}

\item{edge_data}{an optional list of named vectors comprising edge data
attributes. The helper function \code{\link{edge_data}()} is strongly
recommended for use here as it helps bind data specifically to the created
edges.}
}
\description{
Options for node and edge aesthetics and data
}
