% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{node_list_1}
\alias{node_list_1}
\title{Node list - Version 1.}
\format{A data frame with 10 rows and 2 variables:
\describe{
  \item{id}{a unique, monotonically increasing integer ID value}
  \item{label}{a unique label associated with each ID value}
}}
\usage{
node_list_1
}
\description{
A very simple, 2-column data frame that can be used to generate graph nodes.
}
\keyword{datasets}
