% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_present.R
\name{node_present}
\alias{node_present}
\title{Determine whether a specified node is present}
\usage{
node_present(graph, node)
}
\arguments{
\item{graph}{a graph object of class
\code{dgr_graph}.}

\item{node}{either node ID value or a node
label to test for presence in the graph.}
}
\value{
a logical value.
}
\description{
From a graph object of class
\code{dgr_graph}, determine whether a
specified node is present.
}
\examples{
# Create a simple graph with
# a path of four nodes
graph <-
  create_graph() \%>\%
  add_path(
    n = 4,
    type = "path",
    label = c("one", "two",
              "three", "four"))

# Determine if there is a node
# with ID `1` in the graph
node_present(
  graph = graph,
  node = 1)
#> TRUE

# Determine if there is a node
# with ID `5` in the graph
node_present(
  graph = graph,
  node = 5)
#> FALSE

# Determine if there is a node
# with label `two` in the graph
node_present(
  graph = graph,
  node = "two")
#> TRUE
}
