% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/node_type.R
\name{node_type}
\alias{node_type}
\title{Create, read, update, delete, or report status of a node type definition}
\usage{
node_type(graph, node, action = "read", value = NULL)
}
\arguments{
\item{graph}{a graph object of class 'gv_graph' that is created using 'graphviz_graph'.}

\item{node}{a node ID corresponding to the node to be selected.}

\item{action}{the type of operation to perform post-query. To remove the type definition from a node, use either 'delete', 'remove', or 'drop'. To add a type definition to a node with no type set, use 'add' or 'create'. To update a node's type definition, use 'update'. To return the value of a node type, use 'read'. To determine whether there is a type set for the selected node, use 'check'.}

\item{value}{a string denoting the node type, supplied only if either adding or updating a node type definition.}
}
\value{
a graph object of class 'gv_graph'.
}
\description{
From a graph object of class 'gv_graph', query a node in the graph (using the node ID) and perform operations on the type definition for that node.
}

