% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_graph.R
\name{open_graph}
\alias{open_graph}
\title{Read a graph or graph series from disk}
\usage{
open_graph(file)
}
\arguments{
\item{file}{the filename for the graph or graph
series. Optionally, this may contain a path to the
file.}
}
\description{
Load a graph or a graph series object
from disk.
}
