% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/query_graph.R
\name{query_graph}
\alias{query_graph}
\title{Perform query on graph object}
\usage{
query_graph(graph, query)
}
\arguments{
\item{graph}{a graph object of class \code{dgr_graph}.}

\item{query}{a query string}
}
\value{
a graph selection object
}
\description{
With a graph object, perform a query to return a graph selection.
}

