% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_edge_attrs.R
\name{rename_edge_attrs}
\alias{rename_edge_attrs}
\title{Rename an edge attribute}
\usage{
rename_edge_attrs(graph, edge_attr_from, edge_attr_to)
}
\arguments{
\item{graph}{a graph object of class}

\item{edge_attr_from}{the name of the edge attribute
that will be renamed.}

\item{edge_attr_to}{the new name of the edge
attribute column identified in \code{edge_attr_from}.}
}
\value{
a graph object of class \code{dgr_graph}.
}
\description{
Within a graph's internal edge data
frame (edf), rename an existing edge attribute.
}
\examples{
library(magrittr)

# Create a random graph
graph <-
  create_random_graph(
  5, 8, set_seed = 3) \%>\%
  set_edge_attrs("color", "green")

# Get the graph's internal edf to show which
# edge attributes are available
get_edge_df(graph)
#>   from to rel color
#> 1    4  1     green
#> 2    2  4     green
#> 3    4  3     green
#> 4    3  5     green
#> 5    5  1     green
#> 6    4  5     green
#> 7    2  1     green
#> 8    3  2     green

# Rename the `value` node attribute as `weight`
graph <-
  graph \%>\%
  rename_edge_attrs("color", "labelfontcolor")

# Get the graph's internal edf to show that the
# edge attribute had been renamed
get_edge_df(graph)
#>   from to rel labelfontcolor
#> 1    4  1              green
#> 2    2  4              green
#> 3    4  3              green
#> 4    3  5              green
#> 5    5  1              green
#> 6    4  5              green
#> 7    2  1              green
#> 8    3  2              green
}

