% Generated by roxygen2 (4.0.1): do not edit by hand
\name{render_diagram}
\alias{render_diagram}
\title{Render diagram in viewer pane}
\usage{
render_diagram(diagram_code)
}
\arguments{
\item{diagram_code}{either a character object or a path to a text file that contains the code for a diagram.}
}
\description{
This function renders the diagram code in RStudio's viewer pane.
}

