% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_graph_panel.R
\name{render_graph_panel}
\alias{render_graph_panel}
\title{Render several graphs in a single panel}
\usage{
render_graph_panel(...)
}
\arguments{
\item{...}{2 or more graph objects of class
\code{dgr_graph}.}
}
\description{
Using 2 or more \code{dgr_graph}
objects, render theses graph in a single panel.
}
\examples{
\dontrun{
# Create 3 random graphs with increasing
# numbers of edges
graph_1 <-
  create_random_graph(
    n = 25, m = 10,
    set_seed = 23)

graph_2 <-
  create_random_graph(
    n = 25, m = 25,
    set_seed = 23)

graph_3 <-
  create_random_graph(
    n = 25, m = 35,
    set_seed = 23)

# Render all three graphs in the
# order specified
render_graph_panel(
  graph_1, graph_2, graph_3)
}
}
