% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grViz.R
\name{renderGrViz}
\alias{renderGrViz}
\title{Widget render function for use in Shiny}
\usage{
renderGrViz(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{an expression that generates a DiagrammeR graph}

\item{env}{the environment in which to evaluate expr.}

\item{quoted}{is expr a quoted expression (with quote())? This is useful if
you want to save an expression in a variable.}
}
\description{
Widget render function for use in Shiny
}
\seealso{
\code{\link{grVizOutput}} for an example in Shiny
}
