% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_edge_attr_in_selection.R
\name{rescale_edge_attr_in_selection}
\alias{rescale_edge_attr_in_selection}
\title{Rescale numeric edge attribute values for edges in a selection}
\usage{
rescale_edge_attr_in_selection(graph, edge_attr_from, edge_attr_to,
  to_lower_bound = 0, to_upper_bound = 1, from_lower_bound = NULL,
  from_upper_bound = NULL)
}
\arguments{
\item{graph}{a graph object of class \code{dgr_graph} that is created using
\code{create_graph}.}

\item{edge_attr_from}{edge_attr_from}

\item{edge_attr_to}{edge_attr_to}

\item{to_lower_bound}{to_lower_bound}

\item{to_upper_bound}{to_upper_bound}

\item{from_lower_bound}{from_lower_bound}

\item{from_upper_bound}{from_upper_bound}
}
\value{
a graph object of class \code{dgr_graph}.
}
\description{
From a graph object of class \code{dgr_graph}, take a set of
numeric values for edge attributes specified in a selection of edges,
rescale, then write a new set of edge attribute values.
}

