% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_graph.R
\name{save_graph}
\alias{save_graph}
\title{Save a graph or graph series to disk}
\usage{
save_graph(x, file)
}
\arguments{
\item{x}{a graph object of class \code{dgr_graph}
or a graph series object of type \code{dgr_graph_1D}.}

\item{file}{a file name for the graph or
graph series. Provide a character string and the
\code{.dgr} extension will be applied to it.}
}
\description{
Save a graph or a graph series object
to disk.
}
