% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/select_edges.R
\name{select_edges}
\alias{select_edges}
\title{Select edges}
\usage{
select_edges(graph, from = NULL, to = NULL, edge_attr = NULL,
  comparison = NULL, regex = NULL, set_op = "union")
}
\arguments{
\item{graph}{a graph object of class \code{dgr_graph} that is created
using \code{create_graph}.}

\item{from}{an optional vector of node IDs from which the edge is
outgoing for filtering the list of edges present in the graph.}

\item{to}{an optional vector of node IDs to which the edge is
incoming for filtering the list of edges present in the graph.}

\item{edge_attr}{an optional character vector of edge attribute values for
filtering the edges returned.}

\item{comparison}{an optional logical expression for filtering the edges
returned.}

\item{regex}{an optional regular expression (regex) for filtering the
edges returned.}

\item{set_op}{an option to perform a set union, intersection, or
difference on the previous selection, if it exists.}
}
\value{
a graph object of class \code{dgr_graph}.
}
\description{
Select edges from a graph object of class \code{dgr_graph}.
}

