% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/select_graph.R
\name{select_graph}
\alias{select_graph}
\title{Select a graph for a pipeline}
\usage{
select_graph(graph)
}
\arguments{
\item{graph}{a graph of class \code{dgr_graph}}
}
\description{
Select a graph and introduce that graph to
a set of pipeline operations.
}

