% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_last_edge.R
\name{select_last_edge}
\alias{select_last_edge}
\title{Select last edge in a series of edges defined in a graph}
\usage{
select_last_edge(graph)
}
\arguments{
\item{graph}{a graph object of class \code{dgr_graph} that is created
using \code{create_graph}.}
}
\value{
a graph object of class \code{dgr_graph}.
}
\description{
Select the last edge from a graph object of class
\code{dgr_graph}. Strictly, this is the edge definition that
encompasses the last record of the graph's edge data frame. In
practice, this will typically be the last edge created.
}

