% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_nbr_nodes_by_node_attr.R
\name{select_nbr_nodes_by_node_attr}
\alias{select_nbr_nodes_by_node_attr}
\title{Select neighboring nodes based on node attribute
similarity}
\usage{
select_nbr_nodes_by_node_attr(graph, node_attr, match)
}
\arguments{
\item{graph}{a graph object of class
\code{dgr_graph} that is created using
\code{create_graph}.}

\item{node_attr}{the name of the node attribute
to use to compare with adjacent nodes.}

\item{match}{the value of the node attribute to
match on.}
}
\value{
a graph object of class \code{dgr_graph}.
}
\description{
Beginning with a selection of a single
node, select those nodes in a neighborhood
of nodes that have a common node attribute.
}

