% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_nodes_by_degree.R
\name{select_nodes_by_degree}
\alias{select_nodes_by_degree}
\title{Select nodes in the graph based on their degree
values}
\usage{
select_nodes_by_degree(graph, degree_type, degree_values, set_op = "union")
}
\arguments{
\item{graph}{a graph object of class
\code{dgr_graph} that is created using
\code{create_graph}.}

\item{degree_type}{the type of degree, either:
\code{in} for the indegree, \code{out} for the
outdegree, and \code{both} or \code{degree} for the
total degree (i.e., indegree + outdegree).}

\item{degree_values}{a logical expression with a
comparison operator (\code{>}, \code{>=}, \code{<},
\code{<=}, \code{==}, or \code{!=}) followed by a
number for numerical filtering.}

\item{set_op}{the set operation to perform upon
consecutive selections of graph nodes. This can
either be as a \code{union} (the default), as an
intersection of selections with \code{intersect},
or, as a \code{difference} on the previous
selection, if it exists.}
}
\value{
a graph object of class \code{dgr_graph}.
}
\description{
Using a graph object of class
\code{dgr_graph}, create a selection of nodes
that have certain degree values.
}

