% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_cache.R
\name{set_cache}
\alias{set_cache}
\title{Place a vector in graph's cache}
\usage{
set_cache(graph, to_cache, col = NULL)
}
\arguments{
\item{graph}{a graph object of class
\code{dgr_graph} that is created using
\code{create_graph}.}

\item{to_cache}{any vector or data frame.}

\item{col}{if a data frame is provided in
\code{to_cache} then a column name from that data
frame must provided here.}
}
\value{
a graph object of class \code{dgr_graph}.
}
\description{
Place any in the cache of a graph
object of class \code{dgr_graph}.
}
\examples{
library(magrittr)

# Create a random graph
graph <-
  create_random_graph(
    10, 22, set_seed = 1)

# Get the closeness values for all nodes from `1`
# to `10` and store in the graph's cache
graph <-
  graph \%>\%
  set_cache(get_closeness(.), "closeness")

# Get the graph's cache
get_cache(graph)
}

