% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_edge_attr_with_selection.R
\name{set_edge_attr_with_selection}
\alias{set_edge_attr_with_selection}
\title{Set edge attributes based on a selection of edges}
\usage{
set_edge_attr_with_selection(graph, edge_attr, value)
}
\arguments{
\item{graph}{a graph object of class \code{dgr_graph} that is created
using \code{create_graph}.}

\item{edge_attr}{the name of the attribute to set.}

\item{value}{the value to be set for the chosen attribute for the
chosen edges.}
}
\value{
a graph object of class \code{dgr_graph}.
}
\description{
From a graph object of class \code{dgr_graph}, set edge
attribute properties for edges available in a selection.
}

