% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/set_graph_time.R
\name{set_graph_time}
\alias{set_graph_time}
\title{Set graph date-time and timezone}
\usage{
set_graph_time(graph, time = NULL, tz = NULL)
}
\arguments{
\item{graph}{a graph object of class \code{dgr_graph} that is created
using \code{create_graph}.}

\item{time}{the date-time to set for the graph.}

\item{tz}{the timezone to set for the graph.}
}
\value{
a graph object of class \code{dgr_graph}.
}
\description{
Set the time and timezone for a graph object of class
\code{dgr_graph}.
}

