% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/set_node_attr.R
\name{set_node_attr}
\alias{set_node_attr}
\title{Set node attributes}
\usage{
set_node_attr(x, nodes, attr, value)
}
\arguments{
\item{x}{either a graph object of class \code{dgr_graph} that is created
using \code{create_graph}, or a node data frame.}

\item{attr}{the name of the attribute to set.}

\item{value}{the value to be set for the chosen attribute for the
chosen nodes.}

\item{node}{a vector of node IDs.}
}
\value{
either a graph object of class \code{dgr_graph} or a node
data frame, depending on what type of object was supplied to \code{x}.
}
\description{
From a graph object of class \code{dgr_graph} or a node
data frame, set node attribute properties for one or more nodes.
}

