% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_node_attr_w_fcn.R
\name{set_node_attr_w_fcn}
\alias{set_node_attr_w_fcn}
\title{Set node attribute values with a graph function}
\usage{
set_node_attr_w_fcn(graph, fcn, ..., column_name = NULL)
}
\arguments{
\item{graph}{a graph object of class
\code{dgr_graph}.}

\item{fcn}{the name of the function to use.}

\item{...}{arguments and values to pass to
the named function in \code{fcn}, if necessary.}

\item{column_name}{an option to supply a column
name for the new node attribute column. If
\code{NULL} then the column name supplied by the
function will used along with a \code{__A}
suffix.}
}
\value{
either a graph object of class
\code{dgr_graph}.
}
\description{
From a graph object of class
\code{dgr_graph} or a node data frame, set node
attribute properties for all nodes in the graph
using one of several whole-graph functions.
}
\examples{
# Create a random graph
graph <-
  create_random_graph(
    n = 10, m = 22,
    set_seed = 23)

# Get the betweenness values for
# each of the graph's nodes as a
# node attribute
graph_1 <-
  graph \%>\%
  set_node_attr_w_fcn(
    fcn = "get_betweenness")

# Inspect the graph's internal
# node data frame
graph_1 \%>\%
  get_node_df()
#>    id type label value betweenness__A
#> 1   1 <NA>     1   6.0       5.904762
#> 2   2 <NA>     2   2.5       4.904762
#> 3   3 <NA>     3   3.5       1.785714
#> 4   4 <NA>     4   7.5       0.000000
#> 5   5 <NA>     5   8.5       5.738095
#> 6   6 <NA>     6   4.5      20.523810
#> 7   7 <NA>     7  10.0       3.333333
#> 8   8 <NA>     8  10.0       0.000000
#> 9   9 <NA>     9   8.5       3.738095
#> 10 10 <NA>    10  10.0       4.071429

# If a specified function takes argument
# values, these can be supplied as well
graph_2 <-
  graph \%>\%
  set_node_attr_w_fcn(
    fcn = "get_alpha_centrality",
    alpha = 2,
    exo = 2)

# Inspect the graph's internal
# node data frame
graph_2 \%>\%
  get_node_df()
#>    id type label value alpha_centrality__A
#> 1   1 <NA>     1   6.0                   2
#> 2   2 <NA>     2   2.5                   2
#> 3   3 <NA>     3   3.5                   6
#> 4   4 <NA>     4   7.5                   2
#> 5   5 <NA>     5   8.5                  14
#> 6   6 <NA>     6   4.5                  50
#> 7   7 <NA>     7  10.0                  22
#> 8   8 <NA>     8  10.0                 106
#> 9   9 <NA>     9   8.5                 162
#> 10 10 <NA>    10  10.0                 462

graph_3 <-
  graph \%>\%
  set_node_attr_w_fcn(
    fcn = "get_s_connected_cmpts",
    column_name = "scc")

# Inspect the graph's internal
# node data frame
graph_3 \%>\%
  get_node_df()
#>    id type label value scc
#> 1   1 <NA>     1   6.0   4
#> 2   2 <NA>     2   2.5   2
#> 3   3 <NA>     3   3.5   3
#> 4   4 <NA>     4   7.5   1
#> 5   5 <NA>     5   8.5   6
#> 6   6 <NA>     6   4.5   7
#> 7   7 <NA>     7  10.0   5
#> 8   8 <NA>     8  10.0  10
#> 9   9 <NA>     9   8.5   8
#> 10 10 <NA>    10  10.0   9
}
