% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_node_attr_with_selection.R
\name{set_node_attr_with_selection}
\alias{set_node_attr_with_selection}
\title{Set node attributes based on a selection of nodes}
\usage{
set_node_attr_with_selection(graph, node_attr, value)
}
\arguments{
\item{graph}{a graph object of class \code{dgr_graph} that is created
using \code{create_graph}.}

\item{node_attr}{the name of the attribute to set.}

\item{value}{the value to be set for the chosen attribute for the
chosen nodes.}
}
\value{
either a graph object of class \code{dgr_graph}.
}
\description{
From a graph object of class \code{dgr_graph}, set node
attribute properties for nodes available in a selection.
}

