% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_node_attrs_ws.R
\name{set_node_attrs_ws}
\alias{set_node_attrs_ws}
\title{Set node attributes with a node selection}
\usage{
set_node_attrs_ws(graph, node_attr, value)
}
\arguments{
\item{graph}{a graph object of class
\code{dgr_graph} that is created using
\code{create_graph}.}

\item{node_attr}{the name of the attribute to set.}

\item{value}{the value to be set for the chosen
attribute for the nodes in the current selection.}
}
\value{
a graph object of class \code{dgr_graph}.
}
\description{
From a graph object of class
\code{dgr_graph} or a node data frame, set node
attribute properties for nodes present in a node
selection.
}

