% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_node_position.R
\name{set_node_position}
\alias{set_node_position}
\title{Apply a layout position to a single node}
\usage{
set_node_position(graph, node, x, y)
}
\arguments{
\item{graph}{a graph object of class
\code{dgr_graph} that is created using
\code{create_graph}.}

\item{x}{the x coordinate to set for the node.}

\item{y}{the y coordinate to set for the node.}

\item{a}{single-length vector containing a
node ID value for which position information
should be applied.}
}
\value{
a graph object of class \code{dgr_graph}.
}
\description{
Apply position information for a
single node. This is done by setting the \code{x}
and \code{y} attrs for the node ID supplied in
\code{node}.
}

