% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trav_in_edge.R
\name{trav_in_edge}
\alias{trav_in_edge}
\title{Traverse from one or more selected nodes onto adjacent, inward edges}
\usage{
trav_in_edge(graph, rel = NULL)
}
\arguments{
\item{graph}{a graph object of class \code{dgr_graph} that is created
using \code{create_graph}.}

\item{rel}{an optional vector of edge relationship values for filtering
the traversal to inward edges.}
}
\value{
a graph object of class \code{dgr_graph}.
}
\description{
From a graph object of class \code{dgr_graph} move to
incoming edges from a selection of one or more selected nodes, thereby
creating a selection of edges. An optional filter by edge relationship
can limit the set of edges traversed to.
}

