% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trav_out.R
\name{trav_out}
\alias{trav_out}
\title{Traverse outward from a selected node, skipping over edges, and creating
a new node selection}
\usage{
trav_out(graph)
}
\arguments{
\item{graph}{a graph object of class \code{dgr_graph} that is created
using \code{create_graph}.}
}
\value{
a graph object of class \code{dgr_graph}.
}
\description{
From a graph object of class \code{dgr_graph} move outward
from one or more nodes present in a selection to other nodes, replacing
the current nodes in the selection with those nodes traversed to.
}

