% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trav_out_node.R
\name{trav_out_node}
\alias{trav_out_node}
\title{Traverse from one or more selected edges toward adjacent outward nodes}
\usage{
trav_out_node(graph, node_attr = NULL, search = NULL)
}
\arguments{
\item{graph}{a graph object of class \code{dgr_graph} that is created
using \code{create_graph}.}

\item{node_attr}{an optional character vector of node attribute values for
filtering the node ID values returned.}

\item{search}{an option to provide a logical expression with a comparison
operator (\code{>}, \code{<}, \code{==}, or \code{!=}) followed by a number
for numerical filtering, or, a regular expression for filtering the nodes
returned through string matching.}
}
\value{
a graph object of class \code{dgr_graph}.
}
\description{
From a graph object of class \code{dgr_graph} move to
adjacent, outward nodes from a selection of one or more selected edges,
thereby creating a selection of nodes. An optional filter by node type
can limit the set of nodes traversed to.
}

