% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trigger_script.R
\name{trigger_script}
\alias{trigger_script}
\title{Trigger a script embedded in a graph series object}
\usage{
trigger_script(graph_series, script = 1)
}
\arguments{
\item{graph_series}{a graph series object of type \code{dgr_graph_1D}.}

\item{script}{the index of the script character string or path reference held in in the graph series.}
}
\value{
a graph series object of type \code{dgr_graph_1D}.
}
\description{
Run an R script located inside or referenced from the graph series object in order to migrate the state of one or more contained graphs.
}

