% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vivagraph.R
\name{vivagraph}
\alias{vivagraph}
\title{Render graphs with VivaGraphJS}
\usage{
vivagraph(graph = NULL, igrf = NULL, layout = "forceDirected",
  positions = NULL, config = NULL, height = NULL, width = NULL,
  elementId = NULL)
}
\arguments{
\item{igrf}{\code{igraph} graph. We will try to smartly convert the \code{igrf}
using \code{get.data.frame(what = "both")}.}

\item{layout}{\code{string} or a layout function from \code{igraph}.  \code{"forceDirected"}
the default and will use the \code{"forceDirected"} layout engine from
VivaGraphJS/ngraph.forceDirected.  \code{"constant"} is the other \code{string}
option for layout. However, if desired, we can also take advantage of a
layout function from \code{igraph}, by providing this as our \code{layout} parameter.}

\item{positions}{\code{data.frame} of two columns \code{x} and \code{y} with fixed
positions if you intend to provide preset positions for nodes.}

\item{config}{\code{list} of other config options. While currently this does nothing,
we expect to add additional configuration options here.}

\item{height}{\code{string} or \code{integer} with a valid CSS \code{height} for the container
for our htmlwidget.}

\item{width}{\code{string} or \code{integer} with a valid CSS \code{width} for the container
for our htmlwidget.}

\item{nodes_df}{\code{data.frame} or \code{vector} providing an identifier
for each node. If \code{nodes_df} is a \code{data.frame}, then all other
columns besides \code{[,1]} or \code{[,"id"]} will be assigned as attributes.}

\item{edges_df}{\code{data.frame} [or \code{vector} of edges providing a from/source
node identifier and a to/target node identifier.
If \code{edges_df} contains >2 columns, then all other
columns besides \code{[,1:2]} will be assigned as attributes.}

\item{elementID}{\code{string} with a valid CSS \code{id}.}
}
\description{
Render graphs with VivaGraphJS
}
\examples{
\dontrun{
# Create a graph using the \\code{create_nodes}, \\code{create_edges},
# and \\code{create_graph} functions
nodes <-
  create_nodes(nodes = LETTERS,
               type = "letter",
               shape = sample(c("circle", "rectangle"),
                              length(LETTERS),
                              replace = TRUE),
               fillcolor = sample(c("aqua", "gray80",
                                    "pink", "lightgreen",
                                    "azure", "yellow"),
                                  length(LETTERS),
                                  replace = TRUE))

edges <-
  create_edges(from = sample(LETTERS, replace = TRUE),
               to = sample(LETTERS, replace = TRUE),
               relationship = "letter_to_letter")

graph <-
  create_graph(nodes_df = nodes,
               edges_df = edges,
               graph_attrs = "layout = neato",
               node_attrs = c("fontname = Helvetica",
                              "style = filled"),
               edge_attrs = c("color = gray20",
                              "arrowsize = 0.5"))

# Render the graph using the \\code{vivagraph} function
vivagraph(graph$nodes_df, graph$edges_df)

# Using the igraph examples
library(igraph)
library(DiagrammeR)

vivagraph(graph.ring(10))

vivagraph(erdos.renyi.game(100, 1/100))

# Utilizing igraph layouts
vivagraph(erdos.renyi.game(100, 1/100), layout = layout.circle)
vivagraph(graph.famous("Zachary"), layout = layout.grid)
}
}

