% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withdraw_values.R
\name{withdraw_values}
\alias{withdraw_values}
\title{Withdraw a stored vector from a graph object}
\usage{
withdraw_values(graph)
}
\arguments{
\item{graph}{a graph object of class \code{dgr_graph} that is created
using \code{create_graph}.}
}
\value{
a vector of node or edge attribute values.
}
\description{
Get the vector of node or edge attribute values stored
in a graph object of class \code{dgr_graph}.
}

