% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/x11_hex.R
\name{x11_hex}
\alias{x11_hex}
\title{X11 colors and hexadecimal color values
A function to create a data frame with X11 colors and hexadecimal color values}
\usage{
x11_hex()
}
\description{
X11 colors and hexadecimal color values
A function to create a data frame with X11 colors and hexadecimal color values
}

