% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/add_edges.R
\name{add_edges}
\alias{add_edges}
\title{Add edges to an existing graph object}
\usage{
add_edges(graph, edges = NULL, edge_from = NULL, edge_to = NULL,
  type = NULL)
}
\arguments{
\item{graph}{a graph object of class 'gv_graph' that is created using 'graphviz_graph'.}

\item{edges}{an edge data frame that is created using 'create_edges'.}

\item{edge_from}{the outgoing node from which the edge is connected.}

\item{edge_to}{the incoming node to which the edge is connected.}

\item{type}{a string specifying the relationship between the connected nodes.}
}
\value{
a graph object of class 'gv_graph'.
}
\description{
With a graph object of class 'gv_graph', add one or more edges of specified types to nodes within the graph.
}

