% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grViz.R
\name{add_mathjax}
\alias{add_mathjax}
\title{Add MathJax To Your 'grViz' Diagram}
\usage{
add_mathjax(gv = NULL, include_mathjax = TRUE)
}
\arguments{
\item{include_mathjax}{\code{logical} to add mathjax js.  Change
to \code{FALSE} if using with \code{rmarkdown} since
MathJax will likely already be added.}

\item{grViz}{a \code{grViz} htmlwidget}
}
\value{
a \code{grViz} htmlwidget
}
\description{
This is an experimental feature with some known weaknesses, but
we consider the weaknesses worthwhile to add this bit of super
functionality to your diagram.
}

