% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_smallworld_graph.R
\name{add_smallworld_graph}
\alias{add_smallworld_graph}
\title{Add a Watts-Strogatz small-world graph}
\usage{
add_smallworld_graph(graph, dimension, size, neighborhood, p, loops = FALSE,
  multiple = FALSE, set_seed = NULL)
}
\arguments{
\item{graph}{a graph object of class
\code{dgr_graph}.}

\item{dimension}{the dimension of the starting
lattice.}

\item{size}{the size of the lattice across each
dimension.}

\item{neighborhood}{the neighborhood where the
lattice nodes are to be connected.}

\item{p}{the rewiring probability.}

\item{loops}{a logical value (default is
\code{FALSE}) that governs whether loops are
allowed to be created.}

\item{multiple}{a logical value (default is
\code{FALSE}) that governs whether multiple
edges are allowed to be created.}

\item{set_seed}{supplying a value sets a random seed
of the \code{Mersenne-Twister} implementation.}
}
\description{
To an existing graph object, add
a graph built according to the Watts-Strogatz
small-world model, which uses a lattice along
with a rewiring probability to randomly modify
edge definitions.
}
\examples{
# Create an undirected smallworld
# graph with 100 nodes using
# a probability value of 0.05
smallworld_graph <-
  create_graph(
    directed = FALSE) \%>\%
  add_smallworld_graph(
    dimension = 1,
    size = 50,
    neighborhood = 1,
    p = 0.05,
    set_seed = 23)

# Get a count of nodes
smallworld_graph \%>\%
  node_count()
#> [1] 50

# Get a count of edges
smallworld_graph \%>\%
  edge_count()
#> [1] 50
}
