% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_edge_count_ws.R
\name{cache_edge_count_ws}
\alias{cache_edge_count_ws}
\title{Cache a count of edges (available in a
selection) in the graph}
\usage{
cache_edge_count_ws(graph)
}
\arguments{
\item{graph}{a graph object of class
\code{dgr_graph} that is created using
\code{create_graph}.}
}
\value{
a graph object of class \code{dgr_graph}.
}
\description{
From a graph object of class
\code{dgr_graph}, get a count of edges available in
a selection and cache that value in the graph for
later retrieval using \code{get_cache}.
}

