% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_node_attrs_ws.R
\name{cache_node_attrs_ws}
\alias{cache_node_attrs_ws}
\title{Cache node attributes (based on a selection of
nodes) in the graph}
\usage{
cache_node_attrs_ws(graph, node_attr, mode = NULL)
}
\arguments{
\item{graph}{a graph object of class
\code{dgr_graph} that is created using
\code{create_graph}.}

\item{node_attr}{the node attribute from which to
obtain values.}

\item{mode}{a option to recast the returned vector
of node attribute value as \code{numeric} or
\code{character}.}
}
\value{
a graph object of class \code{dgr_graph}.
}
\description{
From a graph object of class
\code{dgr_graph}, get node attribute properties for
nodes available in a selection and cache those
values in the graph for later retrieval using
\code{get_cache}.
}

