% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/combine_edges.R
\name{combine_edges}
\alias{combine_edges}
\title{Combine edge data frames for Graphviz graphs}
\usage{
combine_edges(...)
}
\arguments{
\item{...}{two or more edge data frames, which contain edge IDs and associated attributes}
}
\value{
a combined edge data frame
}
\description{
Combine several edge data frames in the style of rbind, except, it works regardless of the number and ordering of the columns.
}

