% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/combine_graphs.R
\name{combine_graphs}
\alias{combine_graphs}
\title{Combine two graphs into a single graph}
\usage{
combine_graphs(x, y)
}
\arguments{
\item{x}{a \code{DiagrammeR} graph object to which another graph will be joined.}

\item{y}{a \code{DiagrammeR} graph object that is to be joined with the graph
suppled as \code{x}.}
}
\value{
a graph object of class \code{dgr_graph}.
}
\description{
Combine two graphs in order to make a new graph, merging nodes
and edges in the process.
}

