% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/combine_nodes.R
\name{combine_nodes}
\alias{combine_nodes}
\title{Combine node data frames}
\usage{
combine_nodes(df1, df2)
}
\arguments{
\item{df1}{the first data frame containing node IDs and associated attributes}

\item{df2}{the second data frame containing node IDs and associated attributes}
}
\value{
a combined node data frame
}
\description{
Combine two node data frames in the style of rbind, except, it works regardless of the number and ordering of the columns.
}

