% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/combine_nodes.R
\name{combine_nodes}
\alias{combine_nodes}
\title{Combine node data frames for Graphviz graphs}
\usage{
combine_nodes(...)
}
\arguments{
\item{...}{two or more node data frames, which contain node IDs and associated attributes}
}
\value{
a combined node data frame
}
\description{
Combine several node data frames in the style of rbind, except, it works regardless of the number and ordering of the columns.
}

