% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_mutal_nodes.R
\name{count_mutual_nodes}
\alias{count_mutual_nodes}
\title{Get the number mutally-connected node pairs}
\usage{
count_mutual_nodes(graph)
}
\arguments{
\item{graph}{a graph object of class
\code{dgr_graph}.}
}
\value{
a single numeric value representing the
mean distance of the graph.
}
\description{
Get the number mutally-connected
node pairs. This works for directed graphs.
}
\examples{
# Create a cycle graph
graph <-
  create_graph() \%>\%
  add_cycle(n = 5)

# Get a count of mutally-connected
# node pairs
count_mutual_nodes(graph)
#> [1] 0

# Create a full graph and then
# count the mutally-connected
# node pairs
create_graph() \%>\%
  add_full_graph(n = 10) \%>\%
  count_mutual_nodes()
#> [1] 45
}
