% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create_edges.R
\name{create_edges}
\alias{create_edges}
\title{Create edges for Graphviz graphs}
\usage{
create_edges(...)
}
\arguments{
\item{...}{one or more named vectors for edges and associated attributes}
}
\value{
an edge data frame
}
\description{
Combine several named vectors for edges and their attributes.
}

