% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create_edges.R
\name{create_edges}
\alias{create_edges}
\title{Create a data frame with edges and their attributes}
\usage{
create_edges(...)
}
\arguments{
\item{...}{one or more named vectors for edges and associated attributes;
the names for the named vectors must include \code{from} and \code{to}
alongside any named vectors for the edge attributes and ancillary data.}
}
\value{
a data frame.
}
\description{
Combine several named vectors for edges and their attributes
into a data frame, which can be combined with other similarly-generated data
frame, or, added to a graph object.
}
\examples{
\dontrun{
# Create an edge data frame
edges <-
  create_edges(from = c("a", "b", "c"),
               to = c("d", "c", "a"),
               rel = "leading_to")
}
}

