% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create_edges.R
\name{create_edges}
\alias{create_edges}
\title{Create a data frame with edges and their attributes
Combine several named vectors for edges and their attributes into a data frame, which can be combined with other similarly-generated data frame, or, added to a graph object.}
\usage{
create_edges(...)
}
\arguments{
\item{...}{one or more named vectors for edges and associated attributes.}
}
\value{
a data frame.
}
\description{
Create a data frame with edges and their attributes
Combine several named vectors for edges and their attributes into a data frame, which can be combined with other similarly-generated data frame, or, added to a graph object.
}

