% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create_nodes.R
\name{create_nodes}
\alias{create_nodes}
\title{Create nodes for Graphviz graphs}
\usage{
create_nodes(...)
}
\arguments{
\item{...}{one or more named vectors for nodes and associated attributes}
}
\value{
a node data frame
}
\description{
Combine several named vectors for nodes and their attributes.
}

