% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create_nodes.R
\name{create_nodes}
\alias{create_nodes}
\title{Create a data frame with nodes and their attributes}
\usage{
create_nodes(...)
}
\arguments{
\item{...}{one or more named vectors for nodes and associated attributes.}
}
\value{
a data frame.
}
\description{
Combine several named vectors for nodes and their attributes into a data frame, which can be combined with other similarly-generated data frame, or, added to a graph object.
}

