% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create_random_graph.R
\name{create_random_graph}
\alias{create_random_graph}
\title{Create a randomized graph}
\usage{
create_random_graph(n, m)
}
\arguments{
\item{n}{the number of nodes to use in the random graph.}

\item{m}{the number of edges to use in the random graph.}
}
\description{
Create a graph of up to \code{n} nodes with randomized edge
assignments.
}
\examples{
\dontrun{
# Create a random graph with 50 nodes
graph_random <- create_random_graph(50)
}
}

