% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_subgraph_ws.R
\name{create_subgraph_ws}
\alias{create_subgraph_ws}
\title{Create a subgraph based on a selection of nodes
or edges}
\usage{
create_subgraph_ws(graph)
}
\arguments{
\item{graph}{a graph object of class
\code{dgr_graph} that is created using
\code{create_graph}.}
}
\value{
a graph object of class \code{dgr_graph}.
}
\description{
Create a subgraph based on a
selection of nodes or edges extant in the graph
object.
}
\examples{
# Create a simple graph
nodes <-
  create_nodes(
    nodes = 1:6,
    value = c(3.5, 2.6, 9.4, 2.7, 5.2, 2.1))

edges <-
  create_edges(
    from = c(1, 2, 4, 5, 2, 6),
    to = c(2, 4, 1, 3, 5, 5))

graph <-
  create_graph(
    nodes_df = nodes,
    edges_df = edges)

get_nodes(graph)
#> [1] 1 2 3 4 5 6

get_edges(graph, return_type = "vector")
#> [1] "1 -> 2" "2 -> 4" "4 -> 1"
#> [4] "5 -> 3" "2 -> 5" "6 -> 5"

# Create a selection of nodes
graph <-
  select_nodes(
    graph = graph,
    node_attr = "value",
    search = "> 3")

# Create a subgraph based on the selection
subgraph <- create_subgraph_ws(graph)

# Check the nodes available in the subgraph
get_nodes(subgraph)
#> [1] 1 3 5

# Check the edges available in the subgraph
get_edges(subgraph, return_type = "vector")
#> [1] "5 -> 3"
}

