% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_xy_graph.R
\name{create_xy_graph}
\alias{create_xy_graph}
\title{Create an (x, y) scatterplot graph}
\usage{
create_xy_graph(..., aspect_ratio = c(1, 1), x_scale = NULL,
  y_scale = NULL, xy_major_steps = NULL, x_name = NULL, y_name = NULL,
  heading = NULL, right_heading = NULL, x_name_location = "inside",
  include_legend = TRUE, legend_offset = c(0, 0), xy_axis_lab_dist = c(0,
  0), xy_axis_tick_width = c(0.1, 0.1), color_axis_ticks = "gray",
  color_axis_labels = "gray", bg_color = "transparent")
}
\arguments{
\item{...}{one or more NDFs containing points
(created through the use of the \code{create_xy_pts()}
function) or EDFs containing lines (created using
the \code{create_xy_lines()} function).}

\item{aspect_ratio}{the aspect ratio of the plot
area.}

\item{x_scale}{the minimum and maximum x value to
display on the x-axis.}

\item{y_scale}{the minimum and maximum y value to
display on the y-axis.}

\item{xy_major_steps}{the number of major steps
across the x and y directions.}

\item{x_name}{the x-axis name.}

\item{y_name}{the y-axis name.}

\item{heading}{a graph heading placed above the plot
area.}

\item{right_heading}{a right-aligned graph heading
placed above the plot area.}

\item{x_name_location}{the location of the x-axis
name, either \code{inside} (default) or \code{outside}.}

\item{xy_axis_lab_dist}{the distances between the x
and y axes and their labels.}

\item{xy_axis_tick_width}{the widths of the x- and
y-axis tick mark}

\item{color_axis_ticks}{the color of the axis tick
marks.}

\item{color_axis_labels}{the color of the axis
labels.}

\item{bg_color}{the color for the background of the
entire viewable area}

\item{an}{option on whether to include a legend.}

\item{the}{x and y offset (from top-right) of the
legend.}
}
\description{
Create an (x, y) scatterplot graph
using an NDF and an EDF for connecting
lines.
}
\examples{
\dontrun{
library(magrittr)
# Create a series of (x, y) data
# points as an NDF
series_01 <-
  create_xy_pts(
    series_label = "series_01",
    x = c(0.0, 1.3, 3.3, 5.2,
          6.2, 7.9, 8.6, 9.7),
    y = c(2.5, 1.6, 2.7, 3.2,
          4.2, 6.9, 8.7, 5.1),
    shape = "circle",
    line_width = 2.5,
    fill_color = "white",
    line_color = "blue")

# Create an EDF with connections
# between (x, y) data points
series_01_lines <-
  create_xy_lines(
    xy_pts = series_01,
    line_color = "blue",
    line_width = 2.5
  )

# Create an (x, y) plot
xy_plot <-
  create_xy_graph(
    series_01,
    series_01_lines,
    x_name = "x title",
    y_name = "y title",
    heading = "The graph name",
    right_heading = "A right-aligned heading")

# View the plot
xy_plot \%>\% render_graph

# Save the graph as a PDF
xy_plot \%>\% export_graph
}
}

