% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_xy_graph.R
\name{create_xy_graph}
\alias{create_xy_graph}
\title{Create an (x, y) scatterplot graph}
\usage{
create_xy_graph(..., aspect_ratio = c(1, 1), x_scale = NULL,
  y_scale = NULL, xy_major_steps = NULL, x_name = NULL, y_name = NULL,
  heading = NULL, footer = NULL, x_name_location = "inside",
  include_legend = TRUE, legend_offset = c(0, 0), xy_axis_lab_dist = c(0,
  0), xy_axis_tick_width = c(0.1, 0.1), color_axis_ticks = "gray",
  color_axis_labels = "gray", bg_color = "transparent")
}
\arguments{
\item{...}{one or more NDFs containing points
(created through the use of the \code{create_xy_pts()}
function) or EDFs containing lines (created using
the \code{create_xy_lines()} function).}

\item{aspect_ratio}{the aspect ratio of the plot
area.}

\item{x_scale}{the minimum and maximum x value to
display on the x-axis. Not providing values will
result in autoscaling in the x direction based on
provided x values.}

\item{y_scale}{the minimum and maximum y value to
display on the y-axis. Not providing values will
result in autoscaling in the y direction based on
provided y values.}

\item{xy_major_steps}{the number of major steps
across the x and y directions. Not providing values
will result in automatically provided major steps.}

\item{x_name}{the x-axis name.}

\item{y_name}{the y-axis name.}

\item{heading}{text to display as a heading above
the plot area.}

\item{footer}{text to display as a footer below the
plot area.}

\item{x_name_location}{the location of the x-axis
name, either \code{inside} (default) or \code{outside}.}

\item{include_legend}{an option for whether to
include a legend.}

\item{legend_offset}{the x and y offset (from
top-right) of the legend.}

\item{xy_axis_lab_dist}{the distances between the x
and y axes and their labels.}

\item{xy_axis_tick_width}{the widths of the x- and
y-axis tick mark}

\item{color_axis_ticks}{the color of the axis tick
marks.}

\item{color_axis_labels}{the color of the axis
labels.}

\item{bg_color}{the color for the background of the
entire viewable area}
}
\description{
Create an (x, y) scatterplot graph
using an NDF and an EDF for connecting
lines.
}
\examples{
\dontrun{
library(DiagrammeRsvg)
library(magrittr)

# Create three groups (Setosa, Versicolor, and Virginica)
# of x,y datapoints for petal length (x) and petal width (y)
setosa <-
  create_xy_pts(
    series_label = "Setosa",
    x = subset(iris, Species == "setosa")$Petal.Length,
    y = subset(iris, Species == "setosa")$Petal.Width,
    line_color = "red")

versicolor <-
  create_xy_pts(
    series_label = "Versicolor",
    x = subset(iris, Species == "versicolor")$Petal.Length,
    y = subset(iris, Species == "versicolor")$Petal.Width,
    line_color = "green")

virginica <-
  create_xy_pts(
    series_label = "Virginica",
    x = subset(iris, Species == "virginica")$Petal.Length,
    y = subset(iris, Species == "virginica")$Petal.Width,
    line_color = "blue")

# Add these xy points to the `create_xy_graph()`
# function and add axis titles and a heading
iris_length_width <-
  create_xy_graph(
    setosa,
    versicolor,
    virginica,
    x_name = "Petal Length",
    y_name = "Petal Width",
    heading = "Iris Dataset",
    legend_offset = c(0, 6),
    bg_color = "white")

# View the graph
iris_length_width \%>\% render_graph

# Export the graph to a PDF file
iris_length_width \%>\% export_graph("iris.pdf")

# Export the graph to a PDF file
iris_length_width \%>\% export_graph("iris.png")
}
}

