% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_xy_graph.R
\name{create_xy_graph}
\alias{create_xy_graph}
\title{Create an (x, y) scatterplot graph}
\usage{
create_xy_graph(series_pts, series_lines, x_min, x_max, x_divisions, y_min,
  y_max, y_divisions, x_axis_lab_dist = 0.3, y_axis_lab_dist = 0.3,
  x_axis_tick_width = 0.1, y_axis_tick_width = 0.1,
  color_axis_ticks = "gray", color_axis_labels = "gray")
}
\arguments{
\item{series_pts}{an NDF containing points for
graphing.}

\item{series_lines}{an EDF containing connecting
lines.}

\item{x_min}{the minimum x value to display.}

\item{x_max}{the maximum x value to display.}

\item{x_divisions}{the number of divisions between
major x-axis marks.}

\item{y_min}{the minimum y value to display.}

\item{y_max}{the maximum y value to display.}

\item{y_divisions}{the number of divisions between
major y-axis marks.}

\item{x_axis_lab_dist}{the distance between the x
axis and its labels.}

\item{y_axis_lab_dist}{the distance between the y
axis and its labels.}

\item{x_axis_tick_width}{the span of the x-axis tick
marks.}

\item{y_axis_tick_width}{the span of the y-axis tick
marks.}

\item{color_axis_ticks}{the color of the axis tick
marks.}

\item{color_axis_labels}{the color of the axis
labels.}
}
\description{
Create an (x, y) scatterplot graph
using an NDF and an EDF for connecting
lines.
}

