% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_nodes_ws.R
\name{delete_nodes_ws}
\alias{delete_nodes_ws}
\title{Delete all selected nodes in a node selection}
\usage{
delete_nodes_ws(graph)
}
\arguments{
\item{graph}{a graph object of class
\code{dgr_graph} that is created using
\code{create_graph}.}
}
\value{
a graph object of class \code{dgr_graph}.
}
\description{
In a graph object of class
\code{dgr_graph}, delete all nodes present in a
selection.
}

