% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/DiagrammeR.R
\name{DiagrammeROutput}
\alias{DiagrammeROutput}
\title{Widget output function for use in Shiny}
\usage{
DiagrammeROutput(outputId, width = "100\%", height = "400px")
}
\arguments{
\item{outputId}{output variable to read from}

\item{width}{a valid CSS unit for the width or a number, which will be coerced to a string and have "px" appended.}

\item{height}{a valid CSS unit for the height or a number, which will be coerced to a string and have "px" appended.}
}
\description{
Widget output function for use in Shiny
}

