% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/display_graph_object.R
\name{display_graph_object}
\alias{display_graph_object}
\title{Update and display graph object}
\usage{
display_graph_object(graph, width = 400)
}
\arguments{
\item{graph}{a \code{dgr_graph} object, created using the \code{create_graph} function.}

\item{width}{the width of the graph representation in pixels.}
}
\description{
Using a \code{dgr_graph} object, update values of counts for nodes, edges, attributes, directed state, and display the schematic in the RStudio Viewer.
}

