% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_metagraph.R
\name{display_metagraph}
\alias{display_metagraph}
\title{Display a property graph's underlying model}
\usage{
display_metagraph(graph)
}
\arguments{
\item{graph}{a graph object of class
\code{dgr_graph}. This graph must fulfill the
condition of being a property graph, otherwise
the function yields an error.}
}
\description{
With a graph object of class
\code{dgr_graph} that is also a property
graph (i.e., all nodes have an assigned
\code{type} value and all edges have an assigned
\code{rel} value), display its metagraph in the
RStudio Viewer. This representation provides all
combinations of edges of different \code{rel}
values to all nodes with distinct \code{type}
values, including any edges to nodes of the same
\code{type} (shown as loops). The precondition
of the graph being a property graph can be
verified by using the \code{is_property_graph()}
function.
}
\examples{
\dontrun{
# Create a randomized property graph with 1000
# nodes and 1350 edges
property_graph <-
  create_random_graph(1000, 1350, set_seed = 23) \%>\%
  select_nodes_by_degree("deg >= 3") \%>\%
  set_node_attrs_ws("type", "a") \%>\%
  clear_selection() \%>\%
  select_nodes_by_degree("deg < 3") \%>\%
  set_node_attrs_ws("type", "b") \%>\%
  clear_selection() \%>\%
  select_nodes_by_degree("deg == 0") \%>\%
  set_node_attrs_ws("type", "c") \%>\%
  set_node_attr_to_display("type") \%>\%
  select_edges_by_node_id(
    get_node_ids(.) \%>\%
      sample(size = 0.15 * length(.) \%>\%floor())) \%>\%
  set_edge_attrs_ws("rel", "r_1") \%>\%
  invert_selection() \%>\%
  set_edge_attrs_ws("rel", "r_2") \%>\%
  clear_selection() \%>\%
  copy_edge_attrs("rel", "label") \%>\%
  add_global_graph_attrs("fontname", "Helvetica", "edge") \%>\%
  add_global_graph_attrs("fontcolor", "gray50", "edge") \%>\%
  add_global_graph_attrs("fontsize", 10, "edge")

# Display this graph's metagraph, or, underlying
# graph model for a property graph
display_metagraph(property_graph)
}
}
