% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_edge_attrs.R
\name{drop_edge_attrs}
\alias{drop_edge_attrs}
\title{Drop an edge attribute column}
\usage{
drop_edge_attrs(graph, edge_attr)
}
\arguments{
\item{graph}{A graph object of class \code{dgr_graph}.}

\item{edge_attr}{The name of the edge attribute column to drop.}
}
\value{
A graph object of class \code{dgr_graph}.
}
\description{
Within a graph's internal edge data frame (edf), remove an existing edge
attribute.
}
\examples{
# Create a random graph using the
# `add_gnm_graph()` function
graph <-
  create_graph() \%>\%
  add_gnm_graph(
    n = 5,
    m = 6,
    set_seed = 23) \%>\%
  set_edge_attrs(
    edge_attr = value,
    values = 3) \%>\%
  mutate_edge_attrs(
    penwidth = value * 2)

# Get the graph's internal
# edf to show which edge
# attributes are available
graph \%>\% get_edge_df()

# Drop the `value` edge
# attribute
graph <-
  graph \%>\%
  drop_edge_attrs(
    edge_attr = value)

# Get the graph's internal
# edf to show that the edge
# attribute `value` had been
# removed
graph \%>\% get_edge_df()

}
