% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/edge_info.R
\name{edge_info}
\alias{edge_info}
\title{Get detailed information on edges}
\usage{
edge_info(graph)
}
\arguments{
\item{graph}{a graph object of class \code{dgr_graph}.}
}
\value{
a data frame containing information specific to each edge within the graph.
}
\description{
Obtain a data frame with detailed information on edges and their interrelationships within a graph.
}

