% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bridging.R
\name{get_bridging}
\alias{get_bridging}
\title{Get bridging scores}
\usage{
get_bridging(graph)
}
\arguments{
\item{graph}{a graph object of class
\code{dgr_graph}.}
}
\value{
a data frame with bridging scores for
each of the nodes.
}
\description{
Get the bridging scores (based on
Valente's Bridging vertex measure) for all nodes
in a graph.
}
\examples{
# Create a random graph using the
# `add_gnm_graph()` function
graph <-
  create_graph() \%>\%
  add_gnm_graph(
    n = 10,
    m = 12,
    set_seed = 23)

# Get the bridging scores for nodes
# in the graph
get_bridging(graph)
#>    id    bridging
#> 1   1 0.018148148
#> 2   2 0.034953704
#> 3   3 0.008024691
#> 4   4 0.020814815
#> 5   5 0.018209877
#> 6   6 0.044444444
#> 7   7 0.046296296
#> 8   8 0.019537037
#> 9   9 0.044444444
#> 10 10 0.007870370

# Add the bridging scores to
# the graph as a node attribute
graph <-
  graph \%>\%
  join_node_attrs(
    df = get_bridging(.))

# Display the graph's node data frame
get_node_df(graph)
#>    id type label    bridging
#> 1   1 <NA>  <NA> 0.018148148
#> 2   2 <NA>  <NA> 0.034953704
#> 3   3 <NA>  <NA> 0.008024691
#> 4   4 <NA>  <NA> 0.020814815
#> 5   5 <NA>  <NA> 0.018209877
#> 6   6 <NA>  <NA> 0.044444444
#> 7   7 <NA>  <NA> 0.046296296
#> 8   8 <NA>  <NA> 0.019537037
#> 9   9 <NA>  <NA> 0.044444444
#> 10 10 <NA>  <NA> 0.007870370
}
