% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_closeness.R
\name{get_closeness}
\alias{get_closeness}
\title{Get closeness centrality values}
\usage{
get_closeness(graph, direction = "all")
}
\arguments{
\item{graph}{a graph object of class
\code{dgr_graph}.}

\item{direction}{using \code{all} (the default), the
search will ignore edge direction while traversing
through the graph. With \code{out}, measurements of
paths will be from a node whereas with \code{in},
measurements of paths will be to a node.}
}
\value{
a data frame with closeness values for
each of the nodes.
}
\description{
Get the closeness centrality values
for all nodes in a graph.
}
\examples{
# Create a random graph using the
# `add_gnm_graph()` function
graph <-
  create_graph() \%>\%
  add_gnm_graph(
    n = 10,
    m = 12,
    set_seed = 23)

# Get closeness values for all nodes
# in the graph
get_closeness(graph)
#>    id  closeness
#> 1   1 0.05263158
#> 2   2 0.06250000
#> 3   3 0.05555556
#> 4   4 0.06666667
#> 5   5 0.06666667
#> 6   6 0.04166667
#> 7   7 0.04347826
#> 8   8 0.05555556
#> 9   9 0.04166667
#> 10 10 0.04545455

# Add the closeness values to
# the graph as a node attribute
graph <-
  graph \%>\%
  join_node_attrs(
    df = get_closeness(.))

# Display the graph's node data frame
get_node_df(graph)
#>    id type label  closeness
#> 1   1 <NA>  <NA> 0.05263158
#> 2   2 <NA>  <NA> 0.06250000
#> 3   3 <NA>  <NA> 0.05555556
#> 4   4 <NA>  <NA> 0.06666667
#> 5   5 <NA>  <NA> 0.06666667
#> 6   6 <NA>  <NA> 0.04166667
#> 7   7 <NA>  <NA> 0.04347826
#> 8   8 <NA>  <NA> 0.05555556
#> 9   9 <NA>  <NA> 0.04166667
#> 10 10 <NA>  <NA> 0.04545455
}
