% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cmty_louvain.R
\name{get_cmty_louvain}
\alias{get_cmty_louvain}
\title{Get community membership by Louvain optimization}
\usage{
get_cmty_louvain(graph)
}
\arguments{
\item{graph}{a graph object of class
\code{dgr_graph}.}
}
\value{
a data frame with group membership
assignments for each of the nodes.
}
\description{
Through the use of multi-level
optimization of a modularity score, obtain the
group membership values for each of the nodes in
the graph.
}
\examples{
# Create a random graph using the
# `add_gnm_graph()` function
graph <-
  create_graph(
    directed = FALSE) \%>\%
  add_gnm_graph(
    n = 10,
    m = 15,
    set_seed = 23)

# Get the group membership values
# for all nodes in the graph
# through the multi-level
# optimization of modularity
# algorithm
get_cmty_louvain(graph)
#>    id louvain_group
#> 1   1             2
#> 2   2             1
#> 3   3             1
#> 4   4             3
#> 5   5             1
#> 6   6             1
#> 7   7             2
#> 8   8             3
#> 9   9             3
#> 10 10             4

# Add the group membership values to the
# graph as a node attribute
graph <-
  graph \%>\%
  join_node_attrs(
    df = get_cmty_louvain(.))

# Display the graph's node data frame
get_node_df(graph)
#>    id type label louvain_group
#> 1   1 <NA>  <NA>             2
#> 2   2 <NA>  <NA>             1
#> 3   3 <NA>  <NA>             1
#> 4   4 <NA>  <NA>             3
#> 5   5 <NA>  <NA>             1
#> 6   6 <NA>  <NA>             1
#> 7   7 <NA>  <NA>             2
#> 8   8 <NA>  <NA>             3
#> 9   9 <NA>  <NA>             3
#> 10 10 <NA>  <NA>             4
}
