% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_connected_components.R
\name{get_connected_components}
\alias{get_connected_components}
\title{Get all nodes associated with connected components}
\usage{
get_connected_components(graph, return_type = "df")
}
\arguments{
\item{graph}{a graph object of class
\code{dgr_graph} that is created using
\code{create_graph}.}

\item{return_type}{using \code{df} (the default), a
data frame containing the component ID and the nodes
associated with each connected component is
provided. With \code{list}, an anologous list object
is provided.}
}
\value{
a list, data frame, or a vector object,
depending on the value given to \code{return_type}.
}
\description{
Determine which nodes in a graph belong
to different connected components (i.e., distinct
sets of nodes with traversable paths to and from
each node in the set).
}

