% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eccentricity.R
\name{get_eccentricity}
\alias{get_eccentricity}
\title{Get node eccentricities}
\usage{
get_eccentricity(graph)
}
\arguments{
\item{graph}{a graph object of class
\code{dgr_graph}.}
}
\value{
a data frame containing eccentricity values
by node ID value.
}
\description{
Get a data frame with node
eccentricity values.
}
\examples{
# Get the eccentricities for all nodes
# in a randomly-created graph
get_eccentricity(
  graph = create_random_graph(
            5, 7, set_seed = 23))
#>   id eccentricity
#> 1  1            2
#> 2  2            3
#> 3  3            2
#> 4  4            1
#> 5  5            0
}
