% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_edge_df.R
\name{get_edge_df}
\alias{get_edge_df}
\title{Get an edge data frame from a graph}
\usage{
get_edge_df(graph)
}
\arguments{
\item{graph}{a graph object of class \code{dgr_graph} that is created
using \code{create_graph}.}
}
\value{
an edge data frame.
}
\description{
From a graph, obtain an edge data frame with all current
edge attributes.
}

