% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_graph_diameter.R
\name{get_graph_diameter}
\alias{get_graph_diameter}
\title{Get the graph diameter}
\usage{
get_graph_diameter(graph)
}
\arguments{
\item{graph}{a graph object of class
\code{dgr_graph}.}
}
\value{
the graph diameter as a single-length
vector.
}
\description{
Get the graph diameter, which is the
maximum eccentricity.
}
\examples{
# Create a random graph
graph <-
  create_random_graph(
    n = 10, m = 22,
    set_seed = 23)

# Get the graph diameter
get_graph_diameter(graph)
#> [1] 4
}
