% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_node_attr.R
\name{get_node_attr}
\alias{get_node_attr}
\title{Get node attributes}
\usage{
get_node_attr(x, nodes = NULL)
}
\arguments{
\item{x}{either a graph object of class \code{dgr_graph} that is created
using \code{create_graph}, or a node data frame.}

\item{node}{an optional vector of node IDs for filtering list of
nodes present in the graph.}
}
\value{
a node data frame.
}
\description{
From a graph object of class \code{dgr_graph} or a node
data frame, get node attribute properties for one or more nodes.
}
\examples{
\dontrun{
# Create a simple graph
nodes <-
  create_nodes(nodes = c("a", "b", "c", "d"),
               type = "letter",
               label = TRUE,
               value = c(3.5, 2.6, 9.4, 2.7))

edges <-
  create_edges(from = c("a", "b", "c"),
               to = c("d", "c", "a"),
               rel = "leading_to",
               color = c("pink", "blue", "red"))

graph <-
  create_graph(nodes_df = nodes,
               edges_df = edges)

# Get node attributes for all nodes in the graph
get_node_attr(x = graph)

# Get node attributes for nodes "a" and "c" in the graph
get_node_attr(x = graph, nodes = c("a", "c"))
}
}

