% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_nodes.R
\name{get_nodes}
\alias{get_nodes}
\title{Get vector of node IDs}
\usage{
get_nodes(x, node_attr = NULL, comparison = NULL, regex = NULL)
}
\arguments{
\item{x}{either a graph object of class \code{dgr_graph} that is created
using \code{create_graph} or a node data frame.}

\item{node_attr}{an optional character vector of node attribute values for
filtering the node ID values returned.}

\item{comparison}{an optional logical expression for filtering the nodes
returned.}

\item{regex}{an optional regular expression (regex) for filtering the
nodes returned.}
}
\value{
a vector of node ID values.
}
\description{
Obtain a vector of node IDs from a graph object or
a node data frame, optionally using a logical expression or regular
expression for filtering.
}
\examples{
\dontrun{
# Before getting node ID values, create a simple graph
nodes <-
  create_nodes(nodes = c("a", "b", "c", "d"),
               type = "letter",
               color = c("red", "green", "grey", "blue"),
               value = c(3.5, 2.6, 9.4, 2.7))

graph <-
  create_graph(nodes_df = nodes)

# Get a vector of all nodes in a graph
get_nodes(graph)
#> [1] "a" "b" "c" "d"

# Get a vector of node ID values from a node data frame
get_nodes(nodes)
#> [1] "a" "b" "c" "d"

# Get a vector of node ID values using a numeric
# comparison (i.e., all nodes with 'value' attribute
# greater than 3)
get_nodes(graph, node_attr = "value",
          comparison = "> 3")
#> [1] "a" "c"

# Get a vector of node ID values using a regex
# pattern (i.e., all nodes with 'color' attribute
# that begins with "gr")
get_nodes(graph, node_attr = "color",
          regex = "^gr")
#> [1] "b" "c"
}
}

