% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_nodes.R
\name{get_nodes}
\alias{get_nodes}
\title{Get vector of node IDs}
\usage{
get_nodes(..., type = NULL)
}
\arguments{
\item{...}{a collection of node data frames, edge data frames, or a single
graph object.}

\item{type}{an optional character vector of \code{type} values to filter the
node ID values returned.}
}
\value{
a vector of node ID values.
}
\description{
Provides information on the node IDs from one or several node
data frames, edge data frames, or graph objects.
}
\examples{
\dontrun{
# Before getting node ID values, create a simple graph
nodes <-
  create_nodes(nodes = c("a", "b", "c", "d"),
               type = "letter",
               label = TRUE,
               value = c(3.5, 2.6, 9.4, 2.7))

edges <-
  create_edges(from = c("a", "b", "c"),
               to = c("d", "c", "a"),
               rel = "leading_to")

graph <-
  create_graph(nodes_df = nodes,
               edges_df = edges)

# Get a vector of all nodes in a graph
get_nodes(graph)
#> [1] "a" "b" "c" "d"

# Get a vector of node ID values from a node data frame
get_nodes(nodes)
#> [1] "a" "b" "c" "d"

# Get a vector of node ID values from an edge data frame
get_nodes(edges)
#> [1] "a" "b" "c" "d"
}
}

