% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_nodes.R
\name{get_nodes}
\alias{get_nodes}
\title{Get vector of node IDs}
\usage{
get_nodes(...)
}
\arguments{
\item{...}{a collection of node data frames, edge data frames, or a single graph object.}
}
\value{
a vector of node ID values.
}
\description{
Provides information on the node IDs from one or several node data frames, edge data frames, or graph objects.
}

