% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_paths.R
\name{get_paths}
\alias{get_paths}
\title{Get paths from a specified node in a directed graph}
\usage{
get_paths(graph, from = NULL, to = NULL, shortest_path = FALSE,
  longest_path = FALSE, distance = NULL)
}
\arguments{
\item{graph}{a graph object of class \code{dgr_graph}.}

\item{from}{the node from which all paths will be determined.}

\item{to}{the node to which all paths will be determined.}

\item{shortest_path}{an option to return paths that are the shortest in the
set of all determined paths.}

\item{longest_path}{an option to return paths that are the longest in the
set of all determined paths.}

\item{distance}{a vector of integer values that specify which of the valid
paths to return when filtering by distance}
}
\value{
a list of paths, sorted by ascending traversal length, comprising
vectors of node IDs in sequence of traversal through the graph
}
\description{
Obtain a list of all possible paths from a given node within
a directed graph
}

