% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_predecessors.R
\name{get_predecessors}
\alias{get_predecessors}
\title{Get node IDs for predecessor nodes to the specified node}
\usage{
get_predecessors(graph, node)
}
\arguments{
\item{graph}{a graph object of class \code{dgr_graph}.}

\item{node}{a node ID for the selected node.}
}
\value{
a vector of node ID values.
}
\description{
Provides a vector of node IDs for all nodes that have a link to the given node.
}

