% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_successors.R
\name{get_successors}
\alias{get_successors}
\title{Get node IDs for successor nodes to the specified node}
\usage{
get_successors(graph, node)
}
\arguments{
\item{graph}{a graph object of class \code{dgr_graph}.}

\item{node}{a node ID for the selected node.}
}
\value{
a vector of node ID values.
}
\description{
Provides a vector of node IDs for all nodes that have a link from the given node.
}

