% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/graph_count.R
\name{graph_count}
\alias{graph_count}
\title{Count graphs in a graph series object}
\usage{
graph_count(graph_series)
}
\arguments{
\item{graph_series}{a graph series object of type \code{dgr_graph_1D}}
}
\value{
a numeric vector representing a count of graphs in a graph series object.
}
\description{
Counts the total number of graphs in a graph series object.
}

