% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/graphviz_nodes_edges_df.R
\name{graphviz_nodes_edges_df}
\alias{graphviz_nodes_edges_df}
\title{Create DOT code from two data frames}
\usage{
graphviz_nodes_edges_df(nodes_df, edges_df, directed = TRUE)
}
\arguments{
\item{nodes_df}{a data frame containing, at minimum, a column called 'nodes' which contains node IDs for the graph. Optionally, additional columns (named as Graphviz node attributes) can be included with values for the named node attribute.}

\item{edges_df}{a data frame containing, at minimum, a column (called 'edge_op') with edge operations as character strings (in the form of '[node_id] -> [node_id]). Alternatively, there may be two columns (called 'edge_from' and 'edge_to') where node IDs are provided. Optionally, additional columns (named as Graphviz edge attributes) can be included with values for the named edge attribute.}

\item{directed}{with TRUE (the default) or FALSE, either directed or undirected edge operations will be generated, respectively.}
}
\description{
A function to generate DOT code from two data frames: one with nodes and the
other with edge operations.
}

