% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grViz.R
\name{grViz}
\alias{grViz}
\title{R + viz.js}
\usage{
grViz(
  diagram = "",
  engine = "dot",
  allow_subst = TRUE,
  options = NULL,
  width = NULL,
  height = NULL,
  envir = parent.frame()
)
}
\arguments{
\item{diagram}{Spec for a diagram as either text, filename string, or file
connection.}

\item{engine}{String for the Graphviz layout engine; can be \code{dot} (default),
\code{neato}, \code{circo}, or \code{twopi}.}

\item{allow_subst}{A boolean that enables/disables substitution
functionality.}

\item{options}{Parameters supplied to the htmlwidgets framework.}

\item{width}{An optional parameter for specifying the width of the resulting
graphic in pixels.}

\item{height}{An optional parameter for specifying the height of the
resulting graphic in pixels.}

\item{envir}{The environment in which substitution functionality takes place.}
}
\value{
An object of class \code{htmlwidget} that will intelligently print itself
into HTML in a variety of contexts including the R console, within R
Markdown documents, and within Shiny output bindings.
}
\description{
Make diagrams in R using \code{viz.js} with infrastructure provided by
\pkg{htmlwidgets}.
}
