% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/image_icon.R
\name{image_icon}
\alias{image_icon}
\title{Icons and their download locations
Create a data frame with image icons and their web addresses.}
\usage{
image_icon(icon_name)
}
\description{
Icons and their download locations
Create a data frame with image icons and their web addresses.
}

