% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_edge_mutual.R
\name{is_edge_mutual}
\alias{is_edge_mutual}
\title{Is the edge mutual with another edge?}
\usage{
is_edge_mutual(graph, edge)
}
\arguments{
\item{graph}{a graph object of class
\code{dgr_graph}.}

\item{edge}{a numeric edge ID value.}
}
\value{
a logical value.
}
\description{
Determines whether an edge
definition has a mutual analogue with the
same node pair.
}
\examples{
# Create a graph that has mutual
# edges across some node pairs
graph <-
  create_graph() \%>\%
  add_path(n = 4) \%>\%
  add_edge(
    from = 4,
    to = 3) \%>\%
  add_edge(
    from = 2,
    to = 1)

# Get the graph's internal
# edge data frame
graph \%>\%
  get_edge_df()
#>   id from to  rel
#> 1  1    1  2 <NA>
#> 2  2    2  3 <NA>
#> 3  3    3  4 <NA>
#> 4  4    4  3 <NA>
#> 5  5    2  1 <NA>

graph \%>\%
  is_edge_mutual(edge = 1)
#> [1] TRUE

graph \%>\%
  is_edge_mutual(edge = 2)
#> [1] FALSE
}
