% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/is_graph_directed.R
\name{is_graph_directed}
\alias{is_graph_directed}
\title{Is the graph a directed graph?}
\usage{
is_graph_directed(graph)
}
\arguments{
\item{graph}{a graph object of class \code{dgr_graph}.}
}
\value{
a logical value.
}
\description{
Determines whether a graph is set to be directed or not and returns a logical value to that effect.
}

