% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/is_graph_empty.R
\name{is_graph_empty}
\alias{is_graph_empty}
\title{Is the graph empty?}
\usage{
is_graph_empty(graph)
}
\arguments{
\item{graph}{a graph object of class \code{dgr_graph}.}
}
\value{
a logical value.
}
\description{
Determines whether a graph contains any nodes and returns a logical value to reflect emptiness.
}

