% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_aes.R
\name{node_aes}
\alias{node_aes}
\title{Insert node aesthetic attributes during node creation}
\usage{
node_aes(shape = NULL, style = NULL, penwidth = NULL, color = NULL,
  fillcolor = NULL, fontname = NULL, fontsize = NULL, fontcolor = NULL,
  height = NULL, width = NULL, x = NULL, y = NULL, group = NULL,
  tooltip = NULL, URL = NULL, distortion = NULL, sides = NULL,
  skew = NULL, peripheries = NULL, gradientangle = NULL,
  fixedsize = NULL, labelloc = NULL, margin = NULL, orientation = NULL)
}
\arguments{
\item{shape}{the shape to use for the node.
Some possible \code{shape} types include:
\code{circle}, \code{rectangle},
\code{triangle}, \code{plaintext},
\code{square}, and \code{polygon}.}

\item{style}{the node line style. The
\code{style} types that can be used are
\code{filled}, \code{invisible},
\code{diagonals}, \code{rounded}, \code{dashed},
\code{dotted}, \code{solid}, and \code{bold}.}

\item{penwidth}{the thickness of the stroke
line for the node shape.}

\item{color}{the color of the node's
outline. Can be an X11 color or a hexadecimal
color code.}

\item{fillcolor}{the color with which
to fill the shape of the node. Can be an X11
color or a hexadecimal color code.}

\item{fontname}{the name of the
system font that will be used for any node
text.}

\item{fontsize}{the point size of
the font used for any node text.}

\item{fontcolor}{the color used
for any node text. Can be an X11 color or a
hexadecimal color code.}

\item{height}{the height of the shape.}

\item{width}{the width of the shape.}

\item{x}{the fixed position of the node in
the x direction. Any integer-based or
floating point value will be accepted.}

\item{y}{the fixed position of the node in
the y direction. Any integer-based or
floating point value will be accepted.}

\item{group}{the node group.}

\item{tooltip}{text for a node tooltip.}

\item{URL}{a URL to associate with a node.
Upon rendering the plot, clicking nodes
with any associated URLs will open the
URL in the default browser.}

\item{distortion}{the level of distortion
for the node.}

\item{sides}{when using the shape
\code{polygon}, this value will provide
the number of sides for that polygon.}

\item{skew}{a \code{0-1} value that will
result in the node shape being skewed
to the right (from bottom to top). A
value in the range \code{0} to \code{-1}
will skew the shape to the left.}

\item{peripheries}{the repeated number
of node shapes (of increasing size) to
draw at the node perhipery.}

\item{gradientangle}{the path angle for
the node color fill gradient.}

\item{fixedsize}{fixedsize}

\item{labelloc}{labelloc}

\item{margin}{margin}

\item{orientation}{orientation}
}
\description{
This helper function should be
invoked to provide values for the namesake
\code{node_aes} argument, which is present
in any function where nodes are created.
}
\examples{
# Create a new graph and add
# a path with several node
# aesthetic attributes
graph <-
  create_graph() \%>\%
  add_path(
    n = 3,
    type = "path",
    node_aes = node_aes(
      shape = "circle",
      x = c(1, 3, 2),
      y = c(4, -1, 3)))

# View the graph's internal
# node data frame; the node
# aesthetic attributes have
# been inserted
graph \%>\%
  get_node_df()
#>   id type label  shape x  y
#> 1  1 path     1 circle 1  4
#> 2  2 path     2 circle 3 -1
#> 3  3 path     3 circle 2  3
}
