% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_count.R
\name{node_count}
\alias{node_count}
\title{Get count of all nodes in the graph}
\usage{
node_count(graph)
}
\arguments{
\item{graph}{a graph object of class
\code{dgr_graph}.}
}
\value{
a numeric vector of single length.
}
\description{
From a graph object of class
\code{dgr_graph}, get a count of nodes in the graph.
}
\examples{
# Set a seed
set.seed(23)

# Create a node data frame (ndf)
ndf <-
  create_node_df(
    n = 26,
    label = TRUE,
    type = c(rep("a", 7),
             rep("b", 9),
             rep("c", 8),
             rep("d", 2)))

# Create an edge data frame (edf)
edf <-
  create_edge_df(
    from = sample(1:26, replace = TRUE),
    to = sample(1:26, replace = TRUE),
    rel = c(rep("rel_a", 7),
            rep("rel_b", 9),
            rep("rel_c", 8),
            rep("rel_d", 2)))

# Create a graph using the ndf and edf
graph <-
  create_graph(
    nodes_df = ndf,
    edges_df = edf)

# Get a count of nodes in the graph
node_count(graph)
#> 26
}
