% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_count.R
\name{node_count}
\alias{node_count}
\title{Get count of all nodes in the graph}
\usage{
node_count(graph)
}
\arguments{
\item{graph}{a graph object of class
\code{dgr_graph}.}
}
\value{
a numeric vector of single length.
}
\description{
From a graph object of class
\code{dgr_graph}, get a count of nodes in the graph.
}
\examples{
# Create a graph with a
# path of nodes and 3
# unconnected nodes
graph <-
  create_graph() \%>\%
  add_path(n = 3) \%>\%
  add_n_nodes(n = 3)

# Get a count of all nodes
# in the graph
graph \%>\%
  node_count()
#> [1] 6
}
