% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/node_count.R
\name{node_count}
\alias{node_count}
\title{Get count of all nodes or certain types of nodes
From a graph object of class \code{dgr_graph}, get a count of nodes in the graph and optionally obtain a count of nodes by their type.}
\usage{
node_count(graph, type = FALSE)
}
\arguments{
\item{graph}{a graph object of class \code{dgr_graph} that is created using \code{create_graph}.}

\item{type}{either a logical value, where \code{TRUE} provides a named vector of node count by type and \code{FALSE} (the default) provides a total count, or, a string corresponding to one or more node types.}
}
\value{
a numeric vector of single length.
}
\description{
Get count of all nodes or certain types of nodes
From a graph object of class \code{dgr_graph}, get a count of nodes in the graph and optionally obtain a count of nodes by their type.
}

