% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/node_info.R
\name{node_info}
\alias{node_info}
\title{Get detailed information on nodes}
\usage{
node_info(graph)
}
\arguments{
\item{graph}{a graph object of class \code{dgr_graph}.}
}
\value{
a data frame containing information specific to each node within the graph.
}
\description{
Obtain a data frame with detailed information on nodes and their interrelationships within a graph.
}

