% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/node_present.R
\name{node_present}
\alias{node_present}
\title{Determine whether a specified node is present in an existing graph object}
\usage{
node_present(graph, node)
}
\arguments{
\item{graph}{a graph object of class \code{dgr_graph} that is created using \code{create_graph}.}

\item{node}{a value that may or may not match a node ID in the graph.}
}
\value{
a logical value.
}
\description{
From a graph object of class \code{dgr_graph}, determine whether a specified node is present.
}

