% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.dgr_graph}
\alias{print.dgr_graph}
\title{Print the graph to the terminal}
\usage{
\method{print}{dgr_graph}(x, ...)
}
\arguments{
\item{x}{a graph object of class \code{dgr_graph}.}
}
\description{
This function will provide a summary of the graph.
}
\examples{
\dontrun{
# Create a random graph using the
# `add_gnm_graph()` function
graph <-
  create_graph() \%>\%
  add_gnm_graph(
    n = 10,
    m = 15,
    set_seed = 23)

# Get a summary of the graph
graph
}
}
\keyword{internal}
