% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/remove_from_graph_series.R
\name{remove_from_graph_series}
\alias{remove_from_graph_series}
\title{Remove graph object from a graph series object
Remove a single graph object from an array of graph objects in a graph series object.}
\usage{
remove_from_graph_series(graph_series, index = "last")
}
\arguments{
\item{graph_series}{a graph series object from which the graph object will be removed.}

\item{index}{the index of the graph object to be removed from the graph series object.}
}
\value{
a graph series object of type 'dgr_graph_1D'.
}
\description{
Remove graph object from a graph series object
Remove a single graph object from an array of graph objects in a graph series object.
}

