% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_edge_attrs.R
\name{rename_edge_attrs}
\alias{rename_edge_attrs}
\title{Rename an edge attribute}
\usage{
rename_edge_attrs(graph, edge_attr_from, edge_attr_to)
}
\arguments{
\item{graph}{a graph object of class \code{dgr_graph}.}

\item{edge_attr_from}{the name of the edge attribute that will be renamed.}

\item{edge_attr_to}{the new name of the edge attribute column identified in
\code{edge_attr_from}.}
}
\value{
a graph object of class \code{dgr_graph}.
}
\description{
Within a graph's internal edge data frame (edf), rename an existing edge
attribute.
}
\examples{
# Create a random graph using the
# `add_gnm_graph()` function
graph <-
  create_graph() \%>\%
  add_gnm_graph(
    n = 5,
    m = 8,
    set_seed = 23) \%>\%
  set_edge_attrs(
    edge_attr = color,
    values = "green")

# Get the graph's internal edf
# to show which edge attributes
# are available
graph \%>\% get_edge_df()

# Rename the `color` node
# attribute as `weight`
graph <-
  graph \%>\%
  rename_edge_attrs(
    edge_attr_from = color,
    edge_attr_to = labelfontcolor)

# Get the graph's internal
# edf to show that the edge
# attribute had been renamed
graph \%>\% get_edge_df()
}
