% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_graph_from_series.R
\name{render_graph_from_series}
\alias{render_graph_from_series}
\title{Render a graph available in a series}
\usage{
render_graph_from_series(graph_series, graph_no, output = "graph",
  width = NULL, height = NULL)
}
\arguments{
\item{graph_series}{a graph series object of type
\code{dgr_graph_1D}.}

\item{graph_no}{the index of the graph in the graph
series.}

\item{output}{a string specifying the output type;
\code{graph} (the default) renders the graph using
the \code{grViz} function, \code{DOT} outputs DOT
code for the graph, and \code{SVG} provides SVG code
for the rendered graph.}

\item{width}{an optional parameter for specifying
the width of the resulting graphic in pixels.}

\item{height}{an optional parameter for specifying
the height of the resulting graphic in pixels.}
}
\description{
Using a graph series object of type
\code{dgr_graph_1D}, either render graph in the
Viewer or output in various formats.
}
\examples{
\dontrun{
# Create three graphs
graph_1 <-
  create_graph() \%>\%
  add_path(n = 4)

graph_2 <-
  create_graph() \%>\%
  add_cycle(n = 5)

graph_3 <-
  create_graph() \%>\%
  add_star(n = 6)

# Create an empty graph series
# and add the graphs
series <-
  create_series() \%>\%
  add_graph_to_graph_series(
    graph = graph_1) \%>\%
  add_graph_to_graph_series(
    graph = graph_2) \%>\%
  add_graph_to_graph_series(
    graph = graph_3)

# View the second graph in
# the series in the Viewer
render_graph_from_series(
  graph_series = series,
  graph_no = 2)
}
}
