% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/select_nodes.R
\name{select_nodes}
\alias{select_nodes}
\title{Select nodes}
\usage{
select_nodes(graph, nodes = NULL, node_attr = NULL, comparison = NULL,
  regex = NULL, set_op = "union")
}
\arguments{
\item{nodes}{an optional vector of node IDs for filtering list of
nodes present in the graph.}

\item{node_attr}{an optional character vector of node attribute values for
filtering the node ID values returned.}

\item{comparison}{an optional logical expression for filtering the nodes
returned.}

\item{regex}{an optional regular expression (regex) for filtering the
nodes returned.}

\item{set_op}{an option to perform a set union, intersection, or
difference on the previous selection, if it exists.}

\item{x}{a graph object of class \code{dgr_graph} that is created
using \code{create_graph}.}
}
\value{
a graph object of class \code{dgr_graph}.
}
\description{
Select nodes from a graph object of class \code{dgr_graph}.
}
\examples{
\dontrun{
# Create a simple graph
nodes <-
  create_nodes(nodes = c("a", "b", "c", "d"),
               type = "letter",
               label = TRUE,
               value = c(3.5, 2.6, 9.4, 2.7))

edges <-
  create_edges(from = c("a", "b", "c"),
               to = c("d", "c", "a"),
               rel = "leading_to")

graph <-
  create_graph(nodes_df = nodes,
               edges_df = edges)

# Set attribute 'color = "green"' for nodes "a" and "c" using
# the graph object
graph <-
  set_node_attr(x = graph, nodes = c("a", "c"),
                attr = "color", value = "green")

# Set attribute 'color = "green"' for nodes "a" and "c" using
# the node data frame
nodes <-
  set_node_attr(x = nodes, nodes = c("a", "c"),
                attr = "color", value = "green")

#' # Set attribute 'color = "blue"' for all nodes
# the node data frame
nodes <-
  set_node_attr(x = nodes, attr = "color", value = "blue")
}
}

