% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/select_nodes.R
\name{select_nodes}
\alias{select_nodes}
\title{Select nodes}
\usage{
select_nodes(graph, nodes = NULL, node_attr = NULL, comparison = NULL,
  regex = NULL, set_op = "union")
}
\arguments{
\item{graph}{a graph object of class \code{dgr_graph} that is created
using \code{create_graph}.}

\item{nodes}{an optional vector of node IDs for filtering list of
nodes present in the graph.}

\item{node_attr}{an optional character vector of node attribute values for
filtering the node ID values returned.}

\item{comparison}{an optional logical expression for filtering the nodes
returned.}

\item{regex}{an optional regular expression (regex) for filtering the
nodes returned.}

\item{set_op}{the set operation to perform upon consecutive selections
of graph nodes. This can either be as a \code{union} (the default), as an
\code{intersection}, or, as a \code{difference} on the previous selection,
if it exists.}
}
\value{
a graph object of class \code{dgr_graph}.
}
\description{
Select nodes from a graph object of class \code{dgr_graph}.
}

