% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/series_info.R
\name{series_info}
\alias{series_info}
\title{Get detailed information on a graph series}
\usage{
series_info(graph_series)
}
\arguments{
\item{graph_series}{a graph series object of type \code{dgr_graph_1D}.}
}
\value{
a data frame containing information on the graphs within the supplied graph series.
}
\description{
Obtain a data frame with detailed information on the graphs within a graph series.
}

