% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/set_edge_attr.R
\name{set_edge_attr}
\alias{set_edge_attr}
\title{Set edge attributes}
\usage{
set_edge_attr(x, from, to, attr, value)
}
\arguments{
\item{x}{either a graph object of class \code{dgr_graph} that is created
using \code{create_graph}, or an edge data frame.}

\item{from}{a vector of node IDs from which the edge is outgoing.}

\item{to}{a vector of node IDs to which the edge is incoming.}

\item{attr}{the name of the attribute to set.}

\item{value}{the value to be set for the chosen attribute for the
chosen edges.}
}
\value{
either a graph object of class \code{dgr_graph} or an edge
data frame, depending on what type of object was supplied to \code{x}.
}
\description{
From a graph object of class \code{dgr_graph} or an edge
data frame, set edge attribute properties for one or more edges
}

