% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_edge_attrs_ws.R
\name{set_edge_attrs_ws}
\alias{set_edge_attrs_ws}
\title{Set edge attributes with an edge selection}
\usage{
set_edge_attrs_ws(graph, edge_attr, value)
}
\arguments{
\item{graph}{a graph object of class
\code{dgr_graph} that is created using
\code{create_graph}.}

\item{edge_attr}{the name of the attribute to set.}

\item{value}{the value to be set for the chosen
attribute for the edges in the current selection.}
}
\value{
a graph object of class \code{dgr_graph}.
}
\description{
From a graph object of class
\code{dgr_graph} or an edge data frame, set edge
attribute properties for one or more edges
}

