% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_graph_name.R
\name{set_graph_name}
\alias{set_graph_name}
\title{Set graph name}
\usage{
set_graph_name(graph, name)
}
\arguments{
\item{graph}{a graph object of class
\code{dgr_graph}.}

\item{name}{the name to set for the graph.}
}
\value{
a graph object of class \code{dgr_graph}.
}
\description{
Set a name for a graph object of class
\code{dgr_graph}.
}
\examples{
# Create an empty graph
graph <- create_graph()

# Provide the new graph with a name
graph <- set_graph_name(graph, "example_name")
}

