% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/subset_series.R
\name{subset_series}
\alias{subset_series}
\title{Subset a graph series object}
\usage{
subset_series(graph_series, by = "number", values, tz = NULL)
}
\arguments{
\item{graph_series}{a graph series object of type \code{dgr_graph_1D}.}

\item{by}{either \code{number}, which allows for subsetting of the graph series by graph indices, or \code{time} which for graph series objects of type \code{temporal} allows for a subsetting of graphs by a date-time or time range.}

\item{values}{where the subsetting of the graph series by to occur via graph indices (where \code{by = number}), provide a vector of those indices; when subsetting by time (where \code{by = time}), a range of times can be provided as a vector.}

\item{tz}{the time zone (\code{tz}) corresponding to dates or date-time string provided in \code{values} (if \code{by = "date"}).}
}
\value{
a graph series object of type \code{dgr_graph_1D}.
}
\description{
Subsetting a graph series by the graphs' index positions in the graph series or through selection via graphs' date-time attributes.
}

