% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/visnetwork.R
\name{visnetwork}
\alias{visnetwork}
\title{Render graph with visNetwork}
\usage{
visnetwork(graph, nav_buttons = TRUE)
}
\arguments{
\item{graph}{a \code{dgr_graph} object, created using the \code{create_graph} function.}

\item{nav_buttons}{a boolean value that determines whether navigation buttons should be displayed along with the rendered graph.}
}
\description{
Render a graph object with the visNetwork R package.
}

