% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/vivagraph.R
\name{vivagraph}
\alias{vivagraph}
\title{Render graphs with VivaGraphJS}
\usage{
vivagraph(network = NULL, options = list(), height = NULL, width = NULL)
}
\arguments{
\item{network}{\code{list} generated from \code{\link{create_graph}}}
}
\description{
Render graphs with VivaGraphJS
}

